Module iota_wallet::account

The account module. Interaction with an Account happens via an AccountHandle.

Re-exports

pub use self::types::OutputDataDto;

Modules

builder 🔒

The module with the AccountBuilder.

constants 🔒

Constants used for the account and account operations.

handle 🔒

A thread guard over an account, all account methods are called from here.

operations 🔒

The account operations like address generation, syncing and creating transactions.

types

Types used in an account and returned from methods.

update 🔒

Methods to update the account state.

Structs

Account

An Account.

AccountHandle

A thread guard over an account, so we can lock the account during operations.

AddressGenerationOptions

Options for address generation

Assets

Features

OutputOptions

StorageDeposit

SyncOptions

The synchronization options

TransactionOptions

Options for transactions

Unlocks

Enums

OutputsToClaim

Enum to specify which outputs should be claimed

RemainderValueStrategy

The strategy to use for the remainder value management when sending funds.