Struct iota_wallet::account::handle::AccountHandle

pub struct AccountHandle {
    account: Arc<RwLock<Account>>,
    pub(crate) client: Client,
    pub(crate) secret_manager: Arc<RwLock<SecretManager>>,
    pub(crate) last_synced: Arc<Mutex<u128>>,
    pub(crate) event_emitter: Arc<Mutex<EventEmitter>>,
    pub(crate) storage_manager: Arc<Mutex<StorageManager>>,
}

A thread guard over an account, so we can lock the account during operations.

Fields
account: Arc<RwLock<Account>>client: Clientsecret_manager: Arc<RwLock<SecretManager>>last_synced: Arc<Mutex<u128>>event_emitter: Arc<Mutex<EventEmitter>>storage_manager: Arc<Mutex<StorageManager>>
Implementations

impl AccountHandle

pub(crate) fn new(
account: Account,
client: Client,
secret_manager: Arc<RwLock<SecretManager>>,
event_emitter: Arc<Mutex<EventEmitter>>,
storage_manager: Arc<Mutex<StorageManager>>
) -> Self

Create a new AccountHandle with an Account

pub async fn alias(&self) -> String

pub async fn get_output(&self, output_id: &OutputId) -> Option<OutputData>

Get the OutputData of an output stored in the account

pub async fn get_foundry_output(
&self,
native_token_id: TokenId
) -> Result<Output>

Get the [Output] that minted a native token by the token ID. First try to get it from the account, if it isn’t in the account try to get it from the node

pub async fn get_transaction(
&self,
transaction_id: &TransactionId
) -> Option<Transaction>

Get the Transaction of a transaction stored in the account

pub async fn get_incoming_transaction_data(
&self,
transaction_id: &TransactionId
) -> Option<(TransactionPayload, Vec<OutputResponse>)>

Get the transaction with inputs of an incoming transaction stored in the account List might not be complete, if the node pruned the data already

pub async fn list_addresses(&self) -> Result<Vec<AccountAddress>>

Returns all addresses of the account

pub async fn list_addresses_with_unspent_outputs(
&self
) -> Result<Vec<AddressWithUnspentOutputs>>

Returns only addresses of the account with balance

pub async fn list_outputs(&self) -> Result<Vec<OutputData>>

Returns all outputs of the account

pub async fn list_unspent_outputs(&self) -> Result<Vec<OutputData>>

Returns all unspent outputs of the account

pub async fn list_transactions(&self) -> Result<Vec<Transaction>>

Returns all transaction of the account

pub async fn list_pending_transactions(&self) -> Result<Vec<Transaction>>

Returns all pending transaction of the account

pub(crate) async fn save(&self, updated_account: Option<&Account>) -> Result<()>

Save the account to the database, accepts the updated_account as option so we don’t need to drop it before saving

impl AccountHandle

pub async fn generate_addresses(
&self,
amount: u32,
options: Option<AddressGenerationOptions>
) -> Result<Vec<AccountAddress>>

Generate addresses and stores them in the account

let public_addresses = account_handle.generate_addresses(2, None).await?;
// internal addresses are used for remainder outputs, if the RemainderValueStrategy for transactions is set to ChangeAddress
let internal_addresses = account_handle
    .generate_addresses(
        1,
        Some(AddressGenerationOptions {
            internal: true,
            ..Default::default()
        }),
    )
    .await?;

pub(crate) async fn generate_remainder_address(&self) -> Result<AccountAddress>

Generate an internal address and store in the account, internal addresses are used for remainder outputs

impl AccountHandle

pub async fn balance(&self) -> Result<AccountBalance>

Get the AccountBalance

impl AccountHandle

pub async fn get_unlockable_outputs_with_additional_unlock_conditions(
&self,
outputs_to_claim: OutputsToClaim
) -> Result<Vec<OutputId>>

Get basic and nft outputs that have [ExpirationUnlockCondition], [StorageDepositReturnUnlockCondition] or [TimelockUnlockCondition] and can be unlocked now and also get basic outputs with only an [AddressUnlockCondition] unlock condition, for additional inputs

pub async fn try_claim_outputs(
&self,
outputs_to_claim: OutputsToClaim
) -> Result<Vec<Transaction>>

Try to claim basic outputs that have additional unlock conditions to their [AddressUnlockCondition].

pub async fn get_basic_outputs_for_additional_inputs(
&self
) -> Result<Vec<OutputData>>

Get basic outputs that have only one unlock condition which is [AddressUnlockCondition], so they can be used as additional inputs

pub async fn claim_outputs(
&self,
output_ids_to_claim: Vec<OutputId>
) -> Result<Vec<Transaction>>

Try to claim basic or nft outputs that have additional unlock conditions to their [AddressUnlockCondition] from AccountHandle::get_unlockable_outputs_with_additional_unlock_conditions().

pub(crate) async fn claim_outputs_internal(
&self,
output_ids_to_claim: Vec<OutputId>,
possible_additional_inputs: Vec<OutputData>
) -> Result<Vec<Transaction>>

Try to claim basic outputs that have additional unlock conditions to their [AddressUnlockCondition].

impl AccountHandle

pub async fn consolidate_outputs(
self: &AccountHandle,
force: bool,
output_consolidation_threshold: Option<usize>
) -> Result<Vec<Transaction>>

Consolidate basic outputs with only an [AddressUnlockCondition] from an account by sending them to the same address again if the output amount is >= the output_consolidation_threshold

impl AccountHandle

pub(crate) async fn search_addresses_with_outputs(
self: &AccountHandle,
address_gap_limit: u32,
sync_options: Option<SyncOptions>
) -> Result<usize>

Search addresses with unspent outputs address_gap_limit defines how many addresses without outputs will be checked in each account, if an address has outputs, the counter is reset Addresses that got crated during this operation and have a higher key_index than the latest one with outputs, will be removed again, to keep the account size smaller

async fn clean_account_after_recovery(
&self,
old_highest_public_address_index: u32,
old_highest_internal_address_index: u32
) -> AccountHandle

During search_addresses_with_outputs we created new addresses that don’t have funds, so we remove them again addresses_len was before we generated new addresses in search_addresses_with_outputs

impl AccountHandle

pub(crate) async fn get_addresses_to_sync(
&self,
options: &SyncOptions
) -> Result<Vec<AddressWithUnspentOutputs>>

Get the addresses that should be synced with the current known unspent output ids Also adds alias and nft addresses from unspent alias or nft outputs that have no Timelock, Expiration or StorageDepositReturn [UnlockCondition]

pub(crate) async fn get_address_output_ids(
&self,
options: &SyncOptions,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>
) -> Result<(Vec<AddressWithUnspentOutputs>, Vec<OutputId>)>

Get the current output ids for provided addresses and only returns addresses that have unspent outputs and return spent outputs separated

pub(crate) async fn get_addresses_outputs(
&self,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>
) -> Result<(Vec<AddressWithUnspentOutputs>, Vec<OutputData>)>

Get outputs from addresses

pub(crate) async fn request_address_output_ids(
&self,
client: &Client,
address: AddressWithUnspentOutputs,
sync_options: &SyncOptions
) -> Result<(AddressWithUnspentOutputs, Vec<OutputId>)>

Returns output ids for outputs that are directly (Ed25519 address in AddressUnlockCondition) or indirectly (alias/nft address in AddressUnlockCondition and the alias/nft output is controlled with the Ed25519 address) connected to

impl AccountHandle

pub(crate) async fn output_response_to_output_data(
&self,
output_responses: Vec<OutputResponse>,
associated_address: &AddressWithUnspentOutputs
) -> Result<Vec<OutputData>>

Convert OutputResponse to OutputData with the network_id added

pub(crate) async fn get_outputs(
&self,
output_ids: Vec<OutputId>,
spent_outputs: bool
) -> Result<(Vec<OutputResponse>, Vec<OutputResponse>)>

Gets outputs by their id, already known outputs are not requested again, but loaded from the account set as unspent, because we wouldn’t get them from the node if they were spent New requested output responses are returned and old ones separated with their accumulated balance

pub(crate) async fn request_incoming_transaction_data(
&self,
transaction_ids: Vec<TransactionId>
) -> Result<()>

impl AccountHandle

pub(crate) async fn sync_pending_transactions(&self) -> Result<()>

Sync transactions and reattach them if unconfirmed. Returns the transaction with updated metadata and spent output ids that don’t need to be locked anymore

impl AccountHandle

pub async fn retry_until_included(
&self,
block_id: &BlockId,
interval: Option<u64>,
max_attempts: Option<u64>
) -> Result<Vec<(BlockId, Block)>>

Retries (promotes or reattaches) a block for provided block id until it’s included (referenced by a milestone). This function is re-exported from the client library and default interval is as defined in iota.rs. Returns the included block at first position and additional reattached blocks

pub async fn sync(&self, options: Option<SyncOptions>) -> Result<AccountBalance>

Sync the account by fetching new information from the nodes. Will also retry pending transactions if necessary.

impl AccountHandle

pub(crate) async fn build_transaction_essence(
&self,
selected_transaction_data: SelectedTransactionData,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to build the transaction essence from the selected in and outputs

impl AccountHandle

pub async fn burn_native_token(
&self,
native_token: (TokenId, U256),
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to burn native tokens. This doesn’t require the foundry output which minted them, but will not increase the foundries melted_tokens field, which makes it impossible to destroy the foundry output.

async fn get_burn_inputs_and_outputs(
&self,
token_id: TokenId,
burn_token_amount: U256
) -> Result<StrippedOutputAggregate>

async fn aggregate_stripped_alias_and_foundry_outputs(
&self,
token_id: TokenId,
burn_token_amount: U256,
stripped_foundry_output: Vec<StrippedOutput>,
stripped_alias_outputs: HashMap<AliasId, StrippedOutput>
) -> Result<StrippedOutputAggregate>

Aggregate foundry and alias outputs that sum up to required burn_token_amount, if controlling alias doesn’t have the specified native token it’s added anyways for unlock purpose

async fn add_controlling_alias_to_aggregate(
&self,
alias_id: AliasId,
aggregate: &mut StrippedOutputAggregate,
stripped_alias_outputs: &mut HashMap<AliasId, StrippedOutput>
) -> Result<bool>

impl AccountHandle

pub async fn burn_nft(
&self,
nft_id: NftId,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to burn an nft output. Outputs controlled by it will be sweeped before if they don’t have a storage deposit return, timelock or expiration unlock condition.

async fn output_id_and_basic_output_for_nft(
&self,
nft_id: NftId
) -> Result<(OutputId, Output)>

impl AccountHandle

pub async fn destroy_alias(
&self,
alias_id: AliasId,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to destroy an alias output. Outputs controlled by it will be sweeped before if they don’t have a storage deposit return, timelock or expiration unlock condition. The amount and possible native tokens will be sent to the governor address.

async fn output_id_and_basic_output_for_alias(
&self,
alias_id: AliasId
) -> Result<(OutputId, Output)>

impl AccountHandle

pub async fn destroy_foundry(
&self,
foundry_id: FoundryId,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to destroy a foundry output with a circulating supply of 0. Native tokens in the foundry (minted by other foundries) will be transactioned to the controlling alias

pub async fn destroy_foundries(
&self,
foundry_ids: HashSet<FoundryId>,
options: Option<TransactionOptions>
) -> Result<Vec<TransactionId>>

Destroy all foundries in the given set foundry_ids

pub(super) async fn find_alias_and_foundry_output_data(
&self,
alias_id: AliasId,
foundry_id: FoundryId
) -> Result<(OutputData, OutputData)>

Find and return unspent OutputData for given alias_id and foundry_id

impl AccountHandle

pub async fn melt_native_token(
&self,
native_token: (TokenId, U256),
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to melt native tokens. This happens with the foundry output which minted them, by increasing it’s melted_tokens field.

impl AccountHandle

pub(crate) async fn get_sweep_remainder_address(
&self,
options: &Option<TransactionOptions>
) -> Result<AddressWrapper>

async fn output_id_and_nft_output(
&self,
nft_id: NftId
) -> Result<(OutputId, NftOutput)>

async fn output_id_and_next_alias_output_state(
&self,
alias_id: AliasId
) -> Result<(OutputId, AliasOutput)>

pub(crate) fn sweep_address_outputs<'a>(
&'a self,
address: Address,
remainder_address: &'a AddressWrapper
) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionId>>> + Send + 'a>>

pub(crate) async fn send_sweep_transaction(
&self,
address: AddressWrapper,
output_ids: impl IntoIterator<Item = OutputId>,
outputs: impl IntoIterator<Item = Output>
) -> Result<TransactionId>

async fn fetch_governor_address_alias_outputs(
&self,
address: &AddressWrapper
) -> Result<Vec<OutputResponse>>

Fetches alias outputs with address set as Governor unlock condition

async fn fetch_address_basic_outputs(
&self,
address: &AddressWrapper
) -> Result<Vec<OutputResponse>>

Fetches basic outputs with address unlock conditions only

async fn fetch_address_nft_outputs(
&self,
address: &AddressWrapper
) -> Result<Vec<OutputResponse>>

Fetches nft outputs with address unlock conditions only

async fn update_unspent_output(
&self,
output_response: OutputResponse,
output_id: OutputId,
network_id: u64
) -> Result<()>

Update unspent output, this function is originally intended for updating recursively synced alias and nft address output

async fn sweep_foundries(
&self,
hrp: &str,
alias_address: &AliasAddress
) -> Result<Vec<TransactionId>>

impl AccountHandle

pub async fn mint_native_token(
&self,
native_token_options: NativeTokenOptions,
options: Option<TransactionOptions>
) -> Result<MintTokenTransaction>

Function to mint native tokens This happens in a two step process:

Create or get an existing alias output

Create a new foundry output with native tokens minted to the account address Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let native_token_options = NativeTokenOptions {
    account_address: None,
    circulating_supply: U256::from(100),
    maximum_supply: U256::from(100),
    foundry_metadata: None
};

let tx = account_handle.mint_native_token(native_token_options, None,).await?;
println!("Transaction created: {}", tx.transaction_id);
if let Some(block_id) = tx.block_id {
    println!("Block sent: {}", block_id);
}

pub(crate) async fn get_or_create_alias_output(
&self,
controller_address: Address,
options: Option<TransactionOptions>
) -> Result<AliasId>

Get an existing alias output or create a new one

impl AccountHandle

pub async fn mint_nfts(
&self,
nfts_options: Vec<NftOptions>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to mint nfts. Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let nft_id: [u8; 38] =
    prefix_hex::decode("08e68f7616cd4948efebc6a77c4f93aed770ac53860100000000000000000000000000000000")?
        .try_into()
        .unwrap();
let nft_options = vec![NftOptions {
    address: Some("rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string()),
    immutable_metadata: Some(b"some immutable nft metadata".to_vec()),
    metadata: Some(b"some nft metadata".to_vec()),
}];

let transaction = account.mint_nfts(nft_options, None).await?;
println!(
    "Transaction: {} Block sent: http://localhost:14265/api/core/v2/blocks/{}",
    transaction.transaction_id,
    transaction.block_id.expect("No block created yet")
);

async fn prepare_mint_nfts(
&self,
nfts_options: Vec<NftOptions>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.mint_nfts()

impl AccountHandle

pub async fn send_amount(
&self,
addresses_with_amount: Vec<AddressWithAmount>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to create basic outputs with which we then will call AccountHandle.send(), the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressWithAmount{
    address: "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string(),
    amount: 1_000_000,
}];

let tx = account_handle.send_amount(outputs, None ).await?;
println!("Transaction created: {}", tx.transaction_id);
if let Some(block_id) = tx.block_id {
    println!("Block sent: {}", block_id);
}

pub async fn prepare_send_amount(
&self,
addresses_with_amount: Vec<AddressWithAmount>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_amount()

impl AccountHandle

pub async fn send_micro_transaction(
&self,
addresses_with_micro_amount: Vec<AddressWithMicroAmount>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to send micro transactions by using the [StorageDepositReturnUnlockCondition] with an [ExpirationUnlockCondition]. Will call AccountHandle.send(), the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressWithMicroAmount{
   address: "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string(),
   amount: 1,
   return_address: None,
   expiration: None,
}];

let transaction = account_handle.send_micro_transaction(outputs, None ).await?;

println!(
   "Transaction: {} Block sent: http://localhost:14265/api/core/v2/blocks/{}",
   transaction.transaction_id,
   transaction.block_id.expect("No block created yet")
);

async fn prepare_send_micro_transaction(
&self,
addresses_with_micro_amount: Vec<AddressWithMicroAmount>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_micro_transaction()

impl AccountHandle

pub async fn send_native_tokens(
&self,
addresses_native_tokens: Vec<AddressNativeTokens>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to send native tokens in basic outputs with a [StorageDepositReturnUnlockCondition] and [ExpirationUnlockCondition], so the storage deposit gets back to the sender and also that the sender gets access to the output again after a defined time (default 1 day), Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressNativeTokens {
    address: "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string(),
    native_tokens: vec![(
        TokenId::from_str("08e68f7616cd4948efebc6a77c4f93aed770ac53860100000000000000000000000000000000")?,
        U256::from(50),
    )],
    ..Default::default()
}];

let tx = account_handle.send_native_tokens(outputs, None).await?;
println!("Transaction created: {}", tx.transaction_id);
if let Some(block_id) = tx.block_id {
    println!("Block sent: {}", block_id);
}

async fn prepare_send_native_tokens(
&self,
addresses_native_tokens: Vec<AddressNativeTokens>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_native_tokens()

impl AccountHandle

pub async fn send_nft(
&self,
addresses_nft_ids: Vec<AddressAndNftId>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Function to send native tokens in basic outputs with a [StorageDepositReturnUnlockCondition](iota_client::block::output::unlock_condition:: StorageDepositReturnUnlockCondition) and [ExpirationUnlockCondition](iota_client::block::output:: unlock_condition::ExpirationUnlockCondition), so the storage deposit gets back to the sender and also that the sender gets access to the output again after a defined time (default 1 day), Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy. Custom inputs will be replaced with the required nft inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressAndNftId {
    address: "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string(),
    nft_id: NftId::from_str("04f9b54d488d2e83a6c90db08ae4b39651bbba8a")?,
}];

let transaction = account.send_nft(outputs, None).await?;

println!(
"Transaction: {} Block sent: http://localhost:14265/api/core/v2/blocks/{}",
transaction.transaction_id,
transaction.block_id.expect("No block created yet")
);

async fn prepare_send_nft(
&self,
addresses_nft_ids: Vec<AddressAndNftId>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_nft()

impl AccountHandle

pub(crate) async fn select_inputs(
&self,
outputs: Vec<Output>,
custom_inputs: Option<Vec<InputSigningData>>,
remainder_address: Option<Address>,
rent_structure: &RentStructure,
allow_burning: bool
) -> Result<SelectedTransactionData>

Selects inputs for a transaction and locks them in the account, so they don’t get used again

impl AccountHandle

pub async fn prepare_output(
&self,
options: OutputOptions,
transaction_options: Option<TransactionOptions>
) -> Result<Output>

Prepare an output for sending If the amount is below the minimum required storage deposit, by default the remaining amount will automatically be added with a StorageDepositReturn UnlockCondition, when setting the ReturnStrategy to gift, the full minimum required storage deposit will be sent to the recipient. When the assets contain an nft_id, the data from the exisiting nft output will be used, just with the address unlock conditions replaced

async fn prepare_nft_output(
&self,
options: OutputOptions,
transaction_options: Option<TransactionOptions>,
nft_id: NftId
) -> Result<Output>

Prepare an nft output

async fn get_remainder_address(
&self,
transaction_options: Option<TransactionOptions>
) -> Result<Address>

impl AccountHandle

pub async fn prepare_transaction(
&self,
outputs: Vec<Output>,
options: Option<TransactionOptions>
) -> Result<PreparedTransactionData>

Get inputs and build the transaction essence

impl AccountHandle

pub async fn sign_transaction_essence(
&self,
prepared_transaction_data: &PreparedTransactionData
) -> Result<SignedTransactionData>

Function to sign a transaction essence

impl AccountHandle

pub(crate) async fn submit_transaction_payload(
&self,
transaction_payload: TransactionPayload
) -> Result<BlockId>

Submits a payload in a block

impl AccountHandle

pub async fn send(
&self,
outputs: Vec<Output>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Send a transaction, if sending a block fails, the function will return None for the block_id, but the wallet will retry sending the transaction during syncing.

let outputs = vec![TransactionOutput {
    address: "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu".to_string(),
    amount: 1_000_000,
    output_kind: None,
}];

let tx = account_handle
    .send(
        outputs,
        Some(TransactionOptions {
            remainder_value_strategy: RemainderValueStrategy::ReuseAddress,
            ..Default::default()
        }),
    )
    .await?;
println!("Transaction created: {}", tx.transaction_id);
if let Some(block_id) = tx.block_id {
    println!("Block sent: {}", block_id);
}

pub async fn finish_transaction(
&self,
outputs: Vec<Output>,
options: Option<TransactionOptions>
) -> Result<Transaction>

Separated function from send, so syncing isn’t called recursively with the consolidation function, which sends transactions

pub async fn sign_and_submit_transaction(
&self,
prepared_transaction_data: PreparedTransactionData
) -> Result<Transaction>

Sign a transaction, submit it to a node and store it in the account

pub async fn submit_and_store_transaction(
&self,
signed_transaction_data: SignedTransactionData
) -> Result<Transaction>

Validate the transaction, submit it to a node and store it in the account

async fn unlock_inputs(&self, inputs: Vec<InputSigningData>) -> Result<()>

impl AccountHandle

pub async fn set_alias(&self, alias: &str) -> Result<()>

pub(crate) async fn update_account(
&self,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>,
unspent_outputs: Vec<OutputData>,
spent_outputs: Vec<OutputId>,
spent_output_responses: Vec<OutputResponse>,
options: &SyncOptions
) -> Result<()>

Update account with newly synced data and emit events for outputs

pub(crate) async fn update_account_with_transactions(
&self,
updated_transactions: Vec<Transaction>,
spent_output_ids: Vec<OutputId>,
output_ids_to_unlock: Vec<OutputId>
) -> Result<()>

Update account with newly synced transactions

pub(crate) async fn update_account_addresses(
&self,
internal: bool,
new_addresses: Vec<AccountAddress>
) -> Result<()>

Update account with newly generated addresses

pub(crate) async fn update_account_with_new_client(
&mut self,
client: Client
) -> Result<()>

pub(crate) async fn update_unspent_outputs(
&self,
output_responses: Vec<OutputResponse>
) -> Result<()>

Update unspent outputs, this function is originally intended for updating recursively synced alias and nft address outputs