Struct iota_wallet::account::Account

pub struct Account {
    id: String,
    signer_type: SignerType,
    index: usize,
    alias: String,
    created_at: DateTime<Local>,
    last_synced_at: Option<DateTime<Local>>,
    addresses: Vec<Address>,
    client_options: ClientOptions,
    storage_path: PathBuf,
    skip_persistence: bool,
    pub(crate) cached_messages: Arc<Mutex<HashMap<MessageId, Message>>>,
}

Account definition.

Fields
id: String

The account identifier.

signer_type: SignerType

The account’s signer type.

index: usize

The account index

alias: String

The account alias.

created_at: DateTime<Local>

Time of account creation.

last_synced_at: Option<DateTime<Local>>

Time the account was last synced with the Tangle.

addresses: Vec<Address>

Address history associated with the seed.

client_options: ClientOptions

The client options.

storage_path: PathBufskip_persistence: boolcached_messages: Arc<Mutex<HashMap<MessageId, Message>>>
Implementations

impl Account

pub fn id(&self) -> &String

The account identifier.

pub fn signer_type(&self) -> &SignerType

The account’s signer type.

pub fn index(&self) -> &usize

The account index

pub fn alias(&self) -> &String

The account alias.

pub fn created_at(&self) -> &DateTime<Local>

Time of account creation.

pub fn last_synced_at(&self) -> &Option<DateTime<Local>>

Time the account was last synced with the Tangle.

pub fn addresses(&self) -> &Vec<Address>

Address history associated with the seed.

pub fn client_options(&self) -> &ClientOptions

The client options.

pub(crate) fn storage_path(&self) -> &PathBuf

pub(crate) fn skip_persistence(&self) -> &bool

pub(crate) fn cached_messages(&self) -> &Arc<Mutex<HashMap<MessageId, Message>>>

impl Account

pub(crate) fn set_id(&mut self, val: String) -> &mut Self

The account identifier.

pub(crate) fn set_last_synced_at(
&mut self,
val: Option<DateTime<Local>>
) -> &mut Self

Time the account was last synced with the Tangle.

pub(crate) fn set_addresses(&mut self, val: Vec<Address>) -> &mut Self

Address history associated with the seed.

pub(crate) fn set_storage_path(&mut self, val: PathBuf) -> &mut Self

pub(crate) fn set_skip_persistence(&mut self, val: bool) -> &mut Self

impl Account

pub(crate) async fn save(&mut self) -> Result<()>

pub(crate) async fn with_messages<T, C: FnOnce(&Vec<MessageIndexation>) -> T>(
&self,
f: C
) -> T

Do something with the indexed messages.

pub fn bech32_hrp(&self) -> String

Returns the address bech32 human readable part.

pub(crate) fn network(&self) -> Network

Returns the address bech32 human readable part.

pub fn latest_address(&self) -> &Address

Returns the most recent address of the account.

pub(crate) fn latest_change_address(&self) -> Option<&Address>

Returns the most recent change address of the account.

fn latest_address_mut(&mut self) -> &mut Address

pub(crate) async fn balance_internal(
&self,
sent_messages: &[Message]
) -> AccountBalance

pub async fn balance(&self) -> Result<AccountBalance>

Gets the account balance information.

pub async fn set_alias(&mut self, alias: impl AsRef<str>) -> Result<()>

Updates the account alias.

pub async fn set_client_options(&mut self, options: ClientOptions) -> Result<()>

Updates the account’s client options.

pub async fn list_messages(
&self,
count: usize,
from: usize,
message_type: Option<MessageType>
) -> Result<Vec<Message>>

Gets a list of transactions on this account. It’s fetched from the storage. To ensure the database is updated with the latest transactions, sync should be called first.

count - Number of (most recent) messages to fetch.

from - Starting point of the subset to fetch.

message_type - Optional message type filter.

Example

use iota_wallet::{
    account_manager::AccountManager, client::ClientOptionsBuilder, message::MessageType, signing::SignerType,
};

#[tokio::main]
async fn main() {
    // gets 10 received messages, skipping the first 5 most recent messages.
    let client_options = ClientOptionsBuilder::new()
        .with_node("https://api.lb-0.h.chrysalis-devnet.iota.cafe")
        .expect("invalid node URL")
        .build()
        .unwrap();
    let mut manager = AccountManager::builder()
        .with_storage("./test-storage", None)
        .unwrap()
        .finish()
        .await
        .unwrap();
    manager.set_stronghold_password("password").await.unwrap();
    manager.store_mnemonic(SignerType::Stronghold, None).await.unwrap();

    let account_handle = manager
        .create_account(client_options)
        .unwrap()
        .initialise()
        .await
        .expect("failed to add account");
    let account = account_handle.read().await;
    account.list_messages(10, 5, Some(MessageType::Received));
}

pub async fn list_spent_addresses(&self) -> Result<Vec<&Address>>

Gets the spent addresses.

pub async fn list_unspent_addresses(&self) -> Result<Vec<&Address>>

Gets the unspent addresses.

pub(crate) fn append_addresses(&mut self, addresses: Vec<Address>)

pub async fn get_node_info(
&self,
url: Option<&str>,
jwt: Option<&str>,
auth: Option<(&str, &str)>
) -> Result<NodeInfoWrapper>

Gets the node info from /api/v1/info endpoint

pub(crate) async fn save_messages(
&mut self,
messages: Vec<Message>
) -> Result<()>

pub async fn get_message(&self, message_id: &MessageId) -> Option<Message>

Gets a message with the given id associated with this account.

pub async fn address_available_balance(&self, address: &Address) -> Result<u64>

Gets the available balance on the given address.