Struct iota_wallet::account::AccountHandle

pub struct AccountHandle {
    account: Arc<RwLock<Account>>,
    pub(crate) client: Client,
    pub(crate) secret_manager: Arc<RwLock<SecretManager>>,
    pub(crate) last_synced: Arc<Mutex<u128>>,
    pub(crate) event_emitter: Arc<Mutex<EventEmitter>>,
    pub(crate) storage_manager: Arc<Mutex<StorageManager>>,
}

A thread guard over an account, so we can lock the account during operations.

Fields
account: Arc<RwLock<Account>>client: Clientsecret_manager: Arc<RwLock<SecretManager>>last_synced: Arc<Mutex<u128>>event_emitter: Arc<Mutex<EventEmitter>>storage_manager: Arc<Mutex<StorageManager>>
Implementations

impl AccountHandle

pub(crate) fn new(
account: Account,
client: Client,
secret_manager: Arc<RwLock<SecretManager>>,
event_emitter: Arc<Mutex<EventEmitter>>,
storage_manager: Arc<Mutex<StorageManager>>
) -> Self

Create a new AccountHandle with an Account

pub async fn alias(&self) -> String

pub async fn get_output(&self, output_id: &OutputId) -> Option<OutputData>

Get the OutputData of an output stored in the account

pub async fn list_addresses(&self) -> Result<Vec<AccountAddress>>

Returns all addresses of the account

pub async fn list_addresses_with_unspent_outputs(
&self
) -> Result<Vec<AddressWithUnspentOutputs>>

Returns only addresses of the account with balance

pub async fn list_outputs(&self) -> Result<Vec<OutputData>>

Returns all outputs of the account

pub async fn list_unspent_outputs(&self) -> Result<Vec<OutputData>>

Returns all unspent outputs of the account

pub async fn list_transactions(&self) -> Result<Vec<Transaction>>

Returns all transaction of the account

pub async fn list_pending_transactions(&self) -> Result<Vec<Transaction>>

Returns all pending transaction of the account

pub(crate) async fn save(&self, updated_account: Option<&Account>) -> Result<()>

Save the account to the database, accepts the updated_account as option so we don’t need to drop it before saving

pub async fn set_alias(&self, alias: &str) -> Result<()>

pub(crate) async fn update_account_with_new_client(
&mut self,
client: Client
) -> Result<()>

impl AccountHandle

pub async fn generate_addresses(
&self,
amount: u32,
options: Option<AddressGenerationOptions>
) -> Result<Vec<AccountAddress>>

Generate addresses and stores them in the account

let public_addresses = account_handle.generate_addresses(2, None).await?;
// internal addresses are used for remainder outputs, if the RemainderValueStrategy for transfers is set to ChangeAddress
let internal_addresses = account_handle
    .generate_addresses(
        1,
        Some(AddressGenerationOptions {
            internal: true,
            ..Default::default()
        }),
    )
    .await?;

impl AccountHandle

pub async fn balance(&self) -> Result<AccountBalance>

Get the AccountBalance

impl AccountHandle

pub(crate) async fn search_addresses_with_funds(
self: &AccountHandle,
address_gap_limit: u32
) -> Result<AccountBalance>

Search addresses with funds address_gap_limit defines how many addresses without balance will be checked in each account, if an address has balance, the counter is reset Addresses that got crated during this operation and have a higher key_index than the latest one with balance, will be removed again, to keep the account size smaller

async fn clean_account_after_recovery(
&self,
old_highest_public_address_index: u32,
old_highest_internal_address_index: u32
) -> AccountHandle

During search_addresses_with_funds we created new addresses that don’t have funds, so we remove them again addresses_len was before we generated new addresses in search_addresses_with_funds

impl AccountHandle

pub async fn get_time_and_milestone_checked(&self) -> Result<(u32, u32)>

Get the local time, but compare it to the time from the nodeinfo, if it’s off more than 5 minutes, an error will be returned

impl AccountHandle

pub async fn get_unlockable_outputs_with_additional_unlock_conditions(
&self,
outputs_to_collect: OutputsToCollect
) -> Result<Vec<OutputId>>

Get basic and nft outputs that have [ExpirationUnlockCondition], [StorageDepositReturnUnlockCondition] or [TimelockUnlockCondition] and can be unlocked now and also get basic outputs with only an [AddressUnlockCondition] unlock condition, for additional inputs

pub async fn try_collect_outputs(
&self,
outputs_to_collect: OutputsToCollect
) -> Result<Vec<TransferResult>>

Try to collect basic outputs that have additional unlock conditions to their [AddressUnlockCondition].

pub async fn get_basic_outputs_for_additional_inputs(
&self
) -> Result<Vec<OutputData>>

Get basic outputs that have only one unlock condition which is [AddressUnlockCondition], so they can be used as additional inputs

pub async fn collect_outputs(
&self,
output_ids_to_collect: Vec<OutputId>
) -> Result<Vec<TransferResult>>

Try to collect basic or nft outputs that have additional unlock conditions to their [AddressUnlockCondition] from AccountHandle::get_unlockable_outputs_with_additional_unlock_conditions().

pub(crate) async fn collect_outputs_internal(
&self,
output_ids_to_collect: Vec<OutputId>,
possible_additional_inputs: Vec<OutputData>
) -> Result<Vec<TransferResult>>

Try to collect basic outputs that have additional unlock conditions to their [AddressUnlockCondition].

impl AccountHandle

pub async fn consolidate_outputs(
self: &AccountHandle,
force: bool
) -> Result<Vec<TransferResult>>

Consolidates basic outputs with only an [AddressUnlockCondition] from an account by sending them to the same address again if the output amount is >= the output_consolidation_threshold

impl AccountHandle

pub(crate) async fn get_addresses_to_sync(
&self,
options: &SyncOptions
) -> Result<Vec<AddressWithUnspentOutputs>>

Get the addresses that should be synced with the current known unspent output ids Also adds alias and nft addresses from unspent alias or nft outputs that have no Timelock, Expiration or StorageDepositReturn [UnlockCondition]

pub(crate) async fn get_address_output_ids(
&self,
options: &SyncOptions,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>
) -> Result<(Vec<AddressWithUnspentOutputs>, Vec<OutputId>)>

Get the current output ids for provided addresses and only returns addresses that have unspent outputs and return spent outputs separated

pub(crate) async fn get_addresses_outputs(
&self,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>
) -> Result<(Vec<AddressWithUnspentOutputs>, Vec<OutputData>)>

Get outputs from addresses

pub(crate) async fn request_address_output_ids(
&self,
client: &Client,
address: AddressWithUnspentOutputs,
sync_options: &SyncOptions
) -> Result<(AddressWithUnspentOutputs, Vec<OutputId>)>

Returns output ids for outputs that are directly (Ed25519 address in AddressUnlockCondition) or indirectly (alias/nft address in AddressUnlockCondition and the alias/nft output is controlled with the Ed25519 address) connected to

impl AccountHandle

pub(crate) async fn output_response_to_output_data(
&self,
output_responses: Vec<OutputResponse>,
associated_address: &AddressWithUnspentOutputs
) -> Result<Vec<OutputData>>

Convert OutputResponse to OutputData with the network_id added

pub(crate) async fn get_outputs(
&self,
output_ids: Vec<OutputId>,
spent_outputs: bool
) -> Result<(Vec<OutputResponse>, u64, Vec<OutputResponse>)>

Gets outputs by their id, already known outputs are not requested again, but loaded from the account set as unspent, because we wouldn’t get them from the node if they were spent New requested output responses are returned and old ones separated with their accumulated balance

impl AccountHandle

pub(crate) async fn sync_pending_transactions(
&self
) -> Result<TransactionSyncResult>

Sync transactions and reattach them if unconfirmed. Returns the transaction with updated metadata and spent output ids that don’t need to be locked anymore

impl AccountHandle

pub async fn sync(&self, options: Option<SyncOptions>) -> Result<AccountBalance>

Syncs the account by fetching new information from the nodes. Will also retry pending transactions and consolidate outputs if necessary.

async fn update_account(
&self,
addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>,
unspent_outputs: Vec<OutputData>,
transaction_sync_result: Option<TransactionSyncResult>,
spent_outputs: Vec<OutputId>,
spent_output_responses: Vec<OutputResponse>,
options: &SyncOptions
) -> Result<()>

Update account with newly synced data

impl AccountHandle

pub(crate) async fn build_transaction_essence(
&self,
selected_transaction_data: SelectedTransactionData,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to build the transaction essence from the selected in and outputs

impl AccountHandle

pub async fn mint_native_token(
&self,
native_token_options: NativeTokenOptions,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to mint native tokens This happens in a two step process:

Create or get an existing alias output

Create a new foundry output with native tokens minted to the account address Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let native_token_options = NativeTokenOptions {
    account_address: None,
    circulating_supply: U256::from(100),
    maximum_supply: U256::from(100),
    foundry_metadata: None
};

let res = account_handle.mint_native_token(native_token_options, None,).await?;
println!("Transaction created: {}", res.1);
if let Some(block_id) = res.0 {
    println!("Block sent: {}", block_id);
}

pub(crate) async fn get_or_create_alias_output(
&self,
controller_address: Address,
options: Option<TransferOptions>
) -> Result<AliasId>

impl AccountHandle

pub async fn mint_nfts(
&self,
nfts_options: Vec<NftOptions>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to mint nfts. Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let nft_id: [u8; 38] =
    hex::decode("08e68f7616cd4948efebc6a77c4f93aed770ac53860100000000000000000000000000000000")?
        .try_into()
        .unwrap();
let nft_options = vec![NftOptions {
    address: Some("atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string()),
    immutable_metadata: Some(b"some immutable nft metadata".to_vec()),
    metadata: Some(b"some nft metadata".to_vec()),
}];

let transfer_result = account.mint_nfts(nft_options, None).await?;
println!(
    "Transaction: {} Block sent: http://localhost:14265/api/v2/blocks/{}",
    transfer_result.transaction_id,
    transfer_result.block_id.expect("No block created yet")
);

pub async fn prepare_mint_nfts(
&self,
nfts_options: Vec<NftOptions>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.mint_nfts()

impl AccountHandle

pub async fn send_amount(
&self,
addresses_with_amount: Vec<AddressWithAmount>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to create basic outputs with which we then will call AccountHandle.send(), the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressWithAmount{
    address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
    amount: 1_000_000,
}];

let res = account_handle.send_amount(outputs, None ).await?;
println!("Transaction created: {}", res.1);
if let Some(block_id) = res.0 {
    println!("Block sent: {}", block_id);
}

pub async fn prepare_send_amount(
&self,
addresses_with_amount: Vec<AddressWithAmount>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_amount()

impl AccountHandle

pub async fn send_micro_transaction(
&self,
addresses_with_micro_amount: Vec<AddressWithMicroAmount>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to send micro transactions by using the [StorageDepositReturnUnlockCondition] with an [ExpirationUnlockCondition]. Will call AccountHandle.send(), the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressWithMicroAmount{
   address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
   amount: 1,
   return_address: None,
   expiration: None,
}];

let transfer_result = account_handle.send_micro_transaction(outputs, None ).await?;

println!(
   "Transaction: {} Block sent: http://localhost:14265/api/v2/blocks/{}",
   transfer_result.transaction_id,
   transfer_result.block_id.expect("No block created yet")
);

pub async fn prepare_send_micro_transaction(
&self,
addresses_with_micro_amount: Vec<AddressWithMicroAmount>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_micro_transaction()

impl AccountHandle

pub async fn send_native_tokens(
&self,
addresses_native_tokens: Vec<AddressNativeTokens>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to send native tokens in basic outputs with a [StorageDepositReturnUnlockCondition] and [ExpirationUnlockCondition], so the storage deposit gets back to the sender and also that the sender gets access to the output again after a defined time (default 1 day), Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy or custom inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressNativeTokens {
    address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
    native_tokens: vec![(
        TokenId::from_str("08e68f7616cd4948efebc6a77c4f93aed770ac53860100000000000000000000000000000000")?,
        U256::from(50),
    )],
    ..Default::default()
}];

let res = account_handle.send_native_tokens(outputs, None).await?;
println!("Transaction created: {}", res.1);
if let Some(block_id) = res.0 {
    println!("Block sent: {}", block_id);
}

pub async fn prepare_send_native_tokens(
&self,
addresses_native_tokens: Vec<AddressNativeTokens>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_native_tokens()

impl AccountHandle

pub async fn send_nft(
&self,
addresses_nft_ids: Vec<AddressAndNftId>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Function to send native tokens in basic outputs with a [StorageDepositReturnUnlockCondition](iota_client::bee_block::output::unlock_condition:: StorageDepositReturnUnlockCondition) and [ExpirationUnlockCondition](iota_client::bee_block::output:: unlock_condition::ExpirationUnlockCondition), so the storage deposit gets back to the sender and also that the sender gets access to the output again after a defined time (default 1 day), Calls AccountHandle.send() internally, the options can define the RemainderValueStrategy. Custom inputs will be replaced with the required nft inputs. Address needs to be Bech32 encoded

let outputs = vec![AddressAndNftId {
    address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
    nft_id: NftId::from_str("04f9b54d488d2e83a6c90db08ae4b39651bbba8a")?,
}];

let transfer_result = account.send_nft(outputs, None).await?;

println!(
"Transaction: {} Block sent: http://localhost:14265/api/v2/blocks/{}",
transfer_result.transaction_id,
transfer_result.block_id.expect("No block created yet")
);

pub async fn prepare_send_nft(
&self,
addresses_nft_ids: Vec<AddressAndNftId>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Function to prepare the transaction for AccountHandle.send_nft()

impl AccountHandle

pub(crate) async fn select_inputs(
&self,
outputs: Vec<Output>,
custom_inputs: Option<Vec<InputSigningData>>,
remainder_address: Option<Address>,
byte_cost_config: &ByteCostConfig
) -> Result<SelectedTransactionData>

Selects inputs for a transaction and locks them in the account, so they don’t get used again

impl AccountHandle

pub async fn prepare_transaction(
&self,
outputs: Vec<Output>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Get inputs and build the transaction essence

impl AccountHandle

pub async fn sign_transaction_essence(
&self,
prepared_transaction_data: &PreparedTransactionData
) -> Result<SignedTransactionData>

Function to sign a transaction essence

impl AccountHandle

pub(crate) async fn submit_transaction_payload(
&self,
transaction_payload: TransactionPayload
) -> Result<BlockId>

Submits a payload in a block

impl AccountHandle

pub async fn send(
&self,
outputs: Vec<Output>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Send a transaction, if sending a block fails, the function will return None for the block_id, but the wallet will retry sending the transaction during syncing.

let outputs = vec![TransferOutput {
    address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
    amount: 1_000_000,
    output_kind: None,
}];

let res = account_handle
    .send(
        outputs,
        Some(TransferOptions {
            remainder_value_strategy: RemainderValueStrategy::ReuseAddress,
            ..Default::default()
        }),
    )
    .await?;
println!("Transaction created: {}", res.1);
if let Some(block_id) = res.0 {
    println!("Block sent: {}", block_id);
}

pub async fn finish_transfer(
&self,
outputs: Vec<Output>,
options: Option<TransferOptions>
) -> Result<TransferResult>

Separated function from send, so syncing isn’t called recursively with the consolidation function, which sends transfers

pub async fn sign_and_submit_transfer(
&self,
prepared_transaction_data: PreparedTransactionData
) -> Result<TransferResult>

Sign a transaction, submit it to a node and store it in the account

pub async fn sync_and_prepare_transaction(
&self,
outputs: Vec<Output>,
options: Option<TransferOptions>
) -> Result<PreparedTransactionData>

Sync an account if not skipped in options and prepare the transaction

pub async fn submit_and_store_transaction(
&self,
signed_transaction_data: SignedTransactionData
) -> Result<TransferResult>

Validate the transaction, submit it to a node and store it in the account

async fn unlock_inputs(&self, inputs: Vec<InputSigningData>) -> Result<()>