Struct iota_wallet::account::AccountInitialiser

pub struct AccountInitialiser {
    accounts: Arc<RwLock<HashMap<String, AccountHandle>>>,
    storage_path: PathBuf,
    alias: Option<String>,
    created_at: Option<DateTime<Local>>,
    messages: Vec<Message>,
    addresses: Vec<Address>,
    client_options: ClientOptions,
    signer_type: Option<SignerType>,
    skip_persistance: bool,
}

Account initialiser.

Fields
accounts: Arc<RwLock<HashMap<String, AccountHandle>>>storage_path: PathBufalias: Option<String>created_at: Option<DateTime<Local>>messages: Vec<Message>addresses: Vec<Address>client_options: ClientOptionssigner_type: Option<SignerType>skip_persistance: bool
Implementations

impl AccountInitialiser

pub(crate) fn new(
client_options: ClientOptions,
accounts: Arc<RwLock<HashMap<String, AccountHandle>>>,
storage_path: PathBuf
) -> Self

Initialises the account builder.

pub fn signer_type(self, signer_type: SignerType) -> Self

Sets the account type.

pub fn alias(self, alias: impl AsRef<str>) -> Self

Defines the account alias. If not defined, we’ll generate one.

pub fn created_at(self, created_at: DateTime<Local>) -> Self

Time of account creation.

pub fn messages(self, messages: Vec<Message>) -> Self

Messages associated with the seed. The account can be initialised with locally stored messages.

pub fn addresses(self, addresses: Vec<Address>) -> Self

The account can be initialised with locally stored address history.

pub fn skip_persistance(self) -> Self

Skips storing the account to the database.

pub async fn initialise(self) -> Result<AccountHandle>

Initialises the account.