Module iota_wallet::account::sync

Modules

input_selection

Structs

AccountSynchronizer
Account sync helper.

SyncedAccount
Data returned from account synchronization.

Enums

AccountSynchronizeStep

RepostAction

Constants

DUST_ALLOWANCE_VALUE

OUTPUT_LOCK_TIMEOUT

Functions

get_address_for_sync

is_dust_allowed

perform_sync

perform_transfer

repost_message

sync_address

sync_addresses
Syncs addresses with the tangle. The method ensures that the wallet local state has all used addresses plus an unused address.

sync_messages
Syncs messages with the tangle. The method should ensures that the wallet local state has messages associated with the address history.