Function iota_wallet::account::sync::sync_addresses

async fn sync_addresses(
    account: &Account, 
    address_index: usize, 
    gap_limit: usize
) -> Result<(Vec<Address>, Vec<(MessageId, Option<bool>, IotaMessage)>)>

Syncs addresses with the tangle. The method ensures that the wallet local state has all used addresses plus an unused address.

To sync addresses for an account from scratch, address_index = 0 and gap_limit = 10 should be provided. To sync addresses from the latest address, address_index = latest address index and gap_limit = 1 should be provided.

Arguments

address_index The address index.

gap_limit Number of addresses indexes that are generated.

Return value

Returns a (addresses, messages) tuples representing the address history up to latest unused address, and the messages associated with the addresses.