Struct iota_wallet::account::sync::AccountSynchronizer

pub struct AccountSynchronizer {
    account_handle: AccountHandle,
    address_index: usize,
    gap_limit: usize,
    skip_persistance: bool,
    steps: Vec<AccountSynchronizeStep>,
}

Account sync helper.

Fields
account_handle: AccountHandleaddress_index: usizegap_limit: usizeskip_persistance: boolsteps: Vec<AccountSynchronizeStep>
Implementations

impl AccountSynchronizer

pub(super) async fn new(account_handle: AccountHandle) -> Self

Initialises a new instance of the sync helper.

pub fn gap_limit(self, limit: usize) -> Self

Number of address indexes that are generated.

pub fn skip_persistance(self) -> Self

Skip saving new messages and addresses on the account object. The found data is returned on the execute call but won’t be persisted on the database.

pub fn address_index(self, address_index: usize) -> Self

Initial address index to start syncing.

pub(crate) fn steps(self, steps: Vec<AccountSynchronizeStep>) -> Self

Sets the steps to run on the sync process. By default it runs all steps (sync_addresses and sync_messages), but the library can pick what to run here.

pub async fn execute(self) -> Result<SyncedAccount>

Syncs account with the tangle. The account syncing process ensures that the latest metadata (balance, transactions) associated with an account is fetched from the tangle and is stored locally.