Struct iota_wallet::account_manager::migration::TritBuf

#[repr(transparent)]
pub(crate) struct TritBuf<T = T1B1Buf<Btrit>>(pub(crate) T)
 where
    T: RawEncodingBuf;

A buffer containing trits.

This type is roughly analogous to Vec or String. It supports pushing and popping trits and dereferences to [Trits]. It may be borrowed as a trit slice, either mutably or immutably.

Tuple Fields
0: T
Implementations

impl<T> TritBuf<T> where
T: RawEncodingBuf,

pub fn new() -> TritBuf<T>

Create a new empty TritBuf.

pub fn with_capacity(cap: usize) -> TritBuf<T>

Create a new empty TritBuf, backed by the given capacity, cap. The resulting TritBuf will contain at least enough space to contain cap trits without needing to reallocate.

pub fn filled(
len: usize,
trit: <<T as RawEncodingBuf>::Slice as RawEncoding>::Trit
) -> TritBuf<T>

Create a new TritBuf of the given length, filled with copies of the provided trit.

pub fn zeros(len: usize) -> TritBuf<T>

Create a new TritBuf of the given length, filled with zero trit.

pub fn from_trits(
trits: &[<<T as RawEncodingBuf>::Slice as RawEncoding>::Trit]
) -> TritBuf<T>

Create a new TritBuf containing the trits from the given slice of trits.

pub fn push(
&mut self,
trit: <<T as RawEncodingBuf>::Slice as RawEncoding>::Trit
)

Push a trit to the back of this TritBuf.

pub fn pop(
&mut self
) -> Option<<<T as RawEncodingBuf>::Slice as RawEncoding>::Trit>

Pop a trit from the back of this TritBuf, returning it if successful.

pub fn append<U>(&mut self, trits: &Trits<U>) where
U: RawEncoding<Trit = <<T as RawEncodingBuf>::Slice as RawEncoding>::Trit> + ?Sized,

Append a trit slice to the end of this TritBuf.

pub fn as_slice(&self) -> &Trits<<T as RawEncodingBuf>::Slice>

Extracts a trit slice containing the data within this buffer.

Note that TritBuf dereferences to Trits anyway, so it’s usually sufficient to take a reference to TritBuf or to just call &Trits methods on it rather than explicitly calling this method first.

pub fn as_slice_mut(&mut self) -> &mut Trits<<T as RawEncodingBuf>::Slice>

Extracts a mutable trit slice containing the data within this buffer.

Note that TritBuf dereferences to Trits anyway, so it’s usually sufficient to take a reference to TritBuf or to just call &mut Trits methods on it rather explicitly calling this method first.

impl TritBuf<T3B1Buf>

pub fn pad_zeros(&mut self)

Pad the trit buffer with [Btrit::Zero] until the buffer’s length is a multiple of 3.

This method is often used in conjunction with [Trites::as_trytes].

pub fn padded_zeros(self) -> TritBuf<T3B1Buf>

Pad the trit buffer with [Btrit::Zero] until the buffer’s length is a multiple of 3.

This method is often used in conjunction with [Trites::as_trytes].

impl<T> TritBuf<T> where
T: RawEncodingBuf,
<T as RawEncodingBuf>::Slice: RawEncoding,
<<T as RawEncodingBuf>::Slice as RawEncoding>::Trit == Btrit,

pub fn from_i8s(
trits: &[i8]
) -> Result<TritBuf<T>, <Btrit as TryFrom<i8>>::Error>

Create a new TritBuf containing the trits given by the slice of i8s.

impl<T> TritBuf<T> where
T: RawEncodingBuf,
<T as RawEncodingBuf>::Slice: RawEncoding,
<<T as RawEncodingBuf>::Slice as RawEncoding>::Trit == Utrit,

pub fn from_u8s(
trits: &[u8]
) -> Result<TritBuf<T>, <Btrit as TryFrom<u8>>::Error>

Create a new TritBuf containing the trits given by the slice of u8s.

impl<T> TritBuf<T1B1Buf<T>> where
T: Trit,
<T as ShiftTernary>::Target: Trit,

pub fn into_shifted(self) -> TritBuf<T1B1Buf<<T as ShiftTernary>::Target>>

Transform this TritBuf into a shifted representation. If the buffer contains balanced trits ([Btrit]), the returned buffer will contain unbalanced trits ([Utrit]).