Struct iota_wallet::account_manager::AccountManager

pub struct AccountManager {
    pub(crate) accounts: Arc<RwLock<Vec<AccountHandle>>>,
    pub(crate) background_syncing_status: Arc<AtomicUsize>,
    pub(crate) client_options: Arc<RwLock<ClientOptions>>,
    pub(crate) coin_type: Arc<AtomicU32>,
    pub(crate) secret_manager: Arc<RwLock<SecretManager>>,
    pub(crate) event_emitter: Arc<Mutex<EventEmitter>>,
    pub(crate) storage_options: StorageOptions,
    pub(crate) storage_manager: Arc<Mutex<StorageManager>>,
}

The account manager, used to create and get accounts. One account manager can hold many accounts, but they should all share the same secret_manager type with the same seed/mnemonic.

Fields
accounts: Arc<RwLock<Vec<AccountHandle>>>background_syncing_status: Arc<AtomicUsize>client_options: Arc<RwLock<ClientOptions>>coin_type: Arc<AtomicU32>secret_manager: Arc<RwLock<SecretManager>>event_emitter: Arc<Mutex<EventEmitter>>storage_options: StorageOptionsstorage_manager: Arc<Mutex<StorageManager>>
Implementations

impl AccountManager

pub async fn recover_accounts(
&self,
account_gap_limit: u32,
address_gap_limit: u32,
sync_options: Option<SyncOptions>
) -> Result<Vec<AccountHandle>>

Find accounts with unspent outputs account_gap_limit defines how many accounts without unspent outputs will be checked, if an account has unspent outputs, the counter is reset address_gap_limit defines how many addresses without unspent outputs will be checked in each account, if an address has unspent outputs, the counter is reset address_start_index and force_syncing will be overwritten in sync_options to not skip addresses, but also don’t send duplicated requests

async fn search_new_accounts(
&self,
account_gap_limit: u32,
address_gap_limit: u32,
max_account_index_to_keep: &mut Option<u32>,
sync_options: Option<SyncOptions>
) -> Result<()>

Generate new accounts and search for unspent outputs

impl AccountManager

pub async fn start_background_syncing(
&self,
options: Option<SyncOptions>,
interval: Option<Duration>
) -> Result<()>

Start the background syncing process for all accounts, default interval is 7 seconds

impl AccountManager

pub async fn get_account<I: Into<AccountIdentifier>>(
&self,
identifier: I
) -> Result<AccountHandle>

Get an account with an AccountIdentifier

impl AccountManager

pub async fn get_ledger_nano_status(&self) -> Result<LedgerNanoStatus>

Get the ledger nano status

impl AccountManager

pub async fn set_stronghold_password(&self, password: &str) -> Result<()>

Sets the Stronghold password

pub async fn change_stronghold_password(
&self,
current_password: &str,
new_password: &str
) -> Result<()>

Change the Stronghold password to another one and also re-encrypt the values in the loaded snapshot with it.

pub async fn set_stronghold_password_clear_interval(
&self,
timeout: Option<Duration>
) -> Result<()>

Sets the Stronghold password clear interval

pub async fn store_mnemonic(&self, mnemonic: String) -> Result<()>

Stores a mnemonic into the Stronghold vault

pub async fn clear_stronghold_password(&self) -> Result<()>

Clears the Stronghold password from memory.

pub async fn is_stronghold_password_available(&self) -> Result<bool>

Checks if the Stronghold password is available.

impl AccountManager

pub async fn backup(
&self,
backup_path: PathBuf,
stronghold_password: String
) -> Result<()>

Backup the account manager data in a Stronghold file stronghold_password must be the current one when Stronghold is used as SecretManager.

pub async fn restore_backup(
&self,
backup_path: PathBuf,
stronghold_password: String
) -> Result<()>

Restore a backup from a Stronghold file Replaces client_options, coin_type, secret_manager and accounts. Returns an error if accounts were already created If Stronghold is used as secret_manager, the existing Stronghold file will be overwritten. If a mnemonic was stored, it will be gone.

impl AccountManager

pub async fn verify_integrity(&self) -> Result<()>

Checks if there is no missing account for example indexes [0, 1, 3] should panic (for now, later return error, automatically fix?) Also checks for each account if there is a gap in an address list and no address is duplicated

impl AccountManager

pub fn builder() -> AccountManagerBuilder

Initialises the account manager builder.

pub fn create_account(&self) -> AccountBuilder

Create a new account

pub async fn get_accounts(&self) -> Result<Vec<AccountHandle>>

Get all accounts

pub async fn remove_latest_account(&self) -> Result<()>

Removes the latest account (account with the largest account index).

pub fn get_secret_manager(&self) -> Arc<RwLock<SecretManager>>

Get the SecretManager

pub async fn set_client_options(&self, options: ClientOptions) -> Result<()>

Sets the client options for all accounts and sets the new bech32_hrp for the addresses.

pub async fn get_client_options(&self) -> ClientOptions

Get the used client options

pub async fn get_node_info(&self) -> Result<NodeInfoWrapper>

Get the node info

pub async fn balance(&self) -> Result<AccountBalance>

Get the balance of all accounts added together

pub async fn sync(&self, options: Option<SyncOptions>) -> Result<AccountBalance>

Sync all accounts

pub fn stop_background_syncing(&self) -> Result<()>

Stop the background syncing of the accounts

pub async fn listen<F>(&self, events: Vec<WalletEventType>, handler: F) where
F: Fn(&Event) + 'static + Clone + Send + Sync,

Listen to wallet events, empty vec will listen to all events

pub async fn clear_listeners(&self, events: Vec<WalletEventType>)

Remove wallet event listeners, empty vec will remove all listeners

pub fn generate_mnemonic(&self) -> Result<String>

Generates a new random mnemonic.

pub fn verify_mnemonic(&self, mnemonic: &str) -> Result<()>

Verify that a &str is a valid mnemonic.

pub async fn emit_test_event(&self, event: WalletEvent) -> Result<()>

Helper function to test events. Emits a provided event with account index 0.

pub async fn delete_accounts_and_database(&self) -> Result<()>

Deletes the accounts and database folder.