Module iota_wallet::actor

The actor interface for the library.

Re-exports

pub use crate::account::AccountIdentifier;

pub use crate::account_manager::AccountManager;

pub use crate::message::Message as WalletMessage;

pub use crate::message::Transfer;

pub use crate::Result;

Modules

message

Structs

AccountDto

The returned account.

AccountToCreate

An account to create.

LegacyAddress

Message

The message type.

MigrationDataDto

Legacy information fetched.

MigrationDataFinder

Finds account data for the migration from legacy network.

MigrationInputDto

Spent address data.

Response

The actor response type.

WalletMessageHandler

The Wallet message handler.

Enums

AccountMethod

Each public account method.

MessageType

The messages that can be sent to the actor.

ResponseType

The response message.

Functions

add_tryte_checksum

Add 9 Trytes checksum to an address and return it as String

convert_async_panics

convert_panics

panic_to_response_message