Enum iota_wallet::actor::message::MessageType

pub enum MessageType {

    RemoveAccount(AccountIdentifier),
    CreateAccount(Box<AccountToCreate>),
    GetAccount(AccountIdentifier),
    GetAccounts,
    CallAccountMethod {
        account_id: AccountIdentifier,
        method: AccountMethod,
    },
    SyncAccounts {
        address_index: Option<usize>,
        gap_limit: Option<usize>,
        account_discovery_threshold: Option<usize>,
    },
    Reattach {
        account_id: AccountIdentifier,
        message_id: String,
    },
    Backup {
        destination: PathBuf,
        password: String,
    },
    RestoreBackup {
        backup_path: String,
        password: String,
    },
    SetStoragePassword(String),
    ClearStoragePassword,
    SetStrongholdPassword(String),
    SetStrongholdPasswordClearInterval(Duration),
    GetStrongholdStatus,
    LockStronghold,
    SendTransfer {
        account_id: AccountIdentifier,
        transfer: Box<TransferBuilder>,
    },
    InternalTransfer {
        from_account_id: AccountIdentifier,
        to_account_id: AccountIdentifier,
        amount: NonZeroU64,
    },
    GenerateMnemonic,
    VerifyMnemonic(String),
    StoreMnemonic {
        signer_type: SignerType,
        mnemonic: Option<String>,
    },
    IsLatestAddressUnused,
    GetLedgerStatus(bool),
    DeleteStorage,
    ChangeStrongholdPassword {
        current_password: String,
        new_password: String,
    },
    SetClientOptions(Box<ClientOptions>),
    GetMigrationData {
        nodes: Vec<String>,
        permanode: Option<String>,
        seed: String,
        security_level: Option<u8>,
        gap_limit: Option<u64>,
        initial_address_index: Option<u64>,
    },
    GetLedgerMigrationData {
        nodes: Vec<String>,
        permanode: Option<String>,
        addresses: Vec<String>,
        security_level: Option<u8>,
    },
    CreateMigrationBundle {
        seed: String,
        input_address_indexes: Vec<u64>,
        mine: bool,
        timeout_secs: u64,
        offset: i64,
        log_file_name: String,
    },
    SendMigrationBundle {
        nodes: Vec<String>,
        bundle_hash: String,
        mwm: u8,
    },
    SendLedgerMigrationBundle {
        nodes: Vec<String>,
        bundle: Vec<String>,
        mwm: u8,
    },
    GetSeedChecksum(String),
    GetMigrationAddress {
        ledger_prompt: bool,
        account_id: AccountIdentifier,
    },
    MineBundle {
        prepared_bundle: Vec<String>,
        spent_bundle_hashes: Vec<String>,
        security_level: u8,
        timeout: u64,
        offset: i64,
    },
    GetLegacyAddressChecksum(String),
    StartBackgroundSync {
        polling_interval: Duration,
        automatic_output_consolidation: bool,
    },
    StopBackgroundSync,
    Participate {
        account_identifier: AccountIdentifier,
        participations: Vec<Participation>,
    },
    StopParticipating {
        account_identifier: AccountIdentifier,
        event_ids: Vec<String>,
    },
    GetParticipationOverview,
    GetParticipationEvents,

}

The messages that can be sent to the actor.

Variants

RemoveAccount(AccountIdentifier)

Tuple Fields

0: AccountIdentifier

Remove the account related to the specified account_id.

CreateAccount(Box<AccountToCreate>)

Tuple Fields

0: Box<AccountToCreate>

Creates an account.

GetAccount(AccountIdentifier)

Tuple Fields

0: AccountIdentifier

Read account.

GetAccounts

Read accounts.

CallAccountMethod

Fields

account_id: AccountIdentifier

The account identifier.

method: AccountMethod

The account method to call.

Consume an account method.

SyncAccounts

Fields

address_index: Option<usize>

The first address index to sync.

gap_limit: Option<usize>

The gap limit.

account_discovery_threshold: Option<usize>

Minimum number of accounts to check on discovery.

Sync accounts.

Reattach

Fields

account_id: AccountIdentifier

The account identifier.

message_id: String

The message to reattach.

Reattach message.

Backup

Fields

destination: PathBuf

The backup destination.

password: String

Stronghold file password.

Backup storage.

RestoreBackup

Fields

backup_path: String

The path to the backed up storage.

password: String

Stronghold file password.

Import accounts from storage.

SetStoragePassword(String)

Tuple Fields

0: String

Sets the password used to encrypt/decrypt the storage.

ClearStoragePassword

Clears the password used to encrypt/decrypt the storage.

SetStrongholdPassword(String)

Tuple Fields

0: String

Set stronghold snapshot password.

SetStrongholdPasswordClearInterval(Duration)

Tuple Fields

0: Duration

Sets the password clear interval.

GetStrongholdStatus

Get stronghold status.

LockStronghold

Lock the stronghold snapshot (clears password and unload snapshot from memory).

SendTransfer

Fields

account_id: AccountIdentifier

The account identifier.

transfer: Box<TransferBuilder>

The transfer details.

Send funds.

InternalTransfer

Fields

from_account_id: AccountIdentifier

The source account identifier.

to_account_id: AccountIdentifier

The destination account identifier.

amount: NonZeroU64

The transfer amount.

Move funds on stored accounts.

GenerateMnemonic

Generates a new mnemonic.

VerifyMnemonic(String)

Tuple Fields

0: String

Checks if the given mnemonic is valid.

StoreMnemonic

Fields

signer_type: SignerType

The signer type.

mnemonic: Option<String>

The mnemonic. If empty, we’ll generate one.

Store mnemonic.

IsLatestAddressUnused

Checks if all accounts has unused latest address after syncing with the Tangle.

GetLedgerStatus(bool)

Tuple Fields

0: bool

Get the Ledger Nano or Speculos simulator status.

DeleteStorage

Deletes the storage.

ChangeStrongholdPassword

Fields

current_password: String

The current stronghold password.

new_password: String

The new stronghold password.

Changes stronghold snapshot password.

SetClientOptions(Box<ClientOptions>)

Tuple Fields

0: Box<ClientOptions>

Updates the client options for all accounts.

GetMigrationData

Fields

nodes: Vec<String>

The nodes to connect to.

permanode: Option<String>

The permanode to use.

seed: String

The legacy seed.

security_level: Option<u8>

The WOTS address security level.

gap_limit: Option<u64>

The gap limit for how many addresses are generated in one round

initial_address_index: Option<u64>

The initial address index.

Get legacy network balance for the seed.

GetLedgerMigrationData

Fields

nodes: Vec<String>

The nodes to connect to.

permanode: Option<String>

The permanode to use.

addresses: Vec<String>

Address objects as String converted with address and index

security_level: Option<u8>

The WOTS address security level.

Get legacy network balance for addresses.

CreateMigrationBundle

Fields

seed: String

The legacy seed.

input_address_indexes: Vec<u64>

The bundle input address indexes.

mine: bool

Whether we should perform bundle mining or not.

timeout_secs: u64

Timeout in seconds for the bundle mining process.

offset: i64

Offset for the bundle mining process.

log_file_name: String

The name of the log file (stored on the storage folder).

Creates the bundle for migration, performs bundle mining if the address was spent and signs the bundle.

SendMigrationBundle

Fields

nodes: Vec<String>

Node URLs.

bundle_hash: String

Bundle hash returned on CreateMigrationBundle.

mwm: u8

Minimum weight magnitude.

Sends the migration bundle associated with the hash.

SendLedgerMigrationBundle

Fields

nodes: Vec<String>

Node URLs.

bundle: Vec<String>

Bundle tx trytes.

mwm: u8

Minimum weight magnitude.

Sends the migration bundle associated with the hash.

GetSeedChecksum(String)

Tuple Fields

0: String

Get seed checksum.

GetMigrationAddress

Fields

ledger_prompt: bool

Display prompt on the ledger

account_id: AccountIdentifier

Account identifier

Get migration address

MineBundle

Fields

prepared_bundle: Vec<String>

Prepared bundle.

spent_bundle_hashes: Vec<String>

Known spent bundle hashes.

security_level: u8

Security level.

timeout: u64

Mining timeout in seconds.

offset: i64

Mining offset.

Mine bundle

GetLegacyAddressChecksum(String)

Tuple Fields

0: String

Get the checksum for a legacy address.

StartBackgroundSync

Fields

polling_interval: Duration

Polling interval, time between each syncing operation

automatic_output_consolidation: bool

If outputs should get automatically consolidated

Start background syncing.

StopBackgroundSync

Stop background syncing.

Participate

Fields

account_identifier: AccountIdentifier

account which should participate

participations: Vec<Participation>

participation information to send

Participate in a voting or staking events

StopParticipating

Fields

account_identifier: AccountIdentifier

account which should stop participate

event_ids: Vec<String>

Event ids to stop participating from

Stop participating in a voting or staking events

GetParticipationOverview

Get participating overview

GetParticipationEvents

Get participation events