Struct iota_wallet::actor::WalletMessageHandler

pub struct WalletMessageHandler {
    account_manager: AccountManager,
}

The Wallet message handler.

Fields
account_manager: AccountManager
Implementations

impl WalletMessageHandler

pub async fn new() -> Result<Self>

Creates a new instance of the message handler with the default account manager.

pub fn with_manager(account_manager: AccountManager) -> Self

Creates a new instance of the message handler with the specified account manager.

pub async fn handle(&self, message: Message)

Handles a message.

async fn backup(
&self,
destination_path: &Path,
password: String
) -> Result<ResponseType>

async fn restore_backup(
&self,
backup_path: &str,
password: String
) -> Result<ResponseType>

async fn reattach(
&self,
account_id: &AccountIdentifier,
message_id: &str
) -> Result<ResponseType>

async fn sync_accounts(
&self,
address_index: &Option<usize>,
gap_limit: &Option<usize>,
account_discovery_threshold: &Option<usize>
) -> Result<ResponseType>

async fn call_account_method(
&self,
account_id: &AccountIdentifier,
method: &AccountMethod
) -> Result<ResponseType>

async fn remove_account(
&self,
account_id: &AccountIdentifier
) -> Result<ResponseType>

The remove account message handler.

async fn create_account(
&self,
account: &AccountToCreate
) -> Result<ResponseType>

The create account message handler.

async fn get_account(
&self,
account_id: &AccountIdentifier
) -> Result<ResponseType>

async fn get_accounts(&self) -> Result<ResponseType>

async fn set_storage_password(&self, password: &str) -> Result<ResponseType>

async fn clear_storage_password(&self) -> Result<ResponseType>

async fn set_stronghold_password(&self, password: &str) -> Result<ResponseType>

async fn send_transfer(
&self,
account_id: &AccountIdentifier,
transfer: Transfer
) -> Result<ResponseType>

async fn internal_transfer(
&self,
from_account_id: &AccountIdentifier,
to_account_id: &AccountIdentifier,
amount: NonZeroU64
) -> Result<ResponseType>