Module iota_wallet::address

The address module.

Structs

Address

An address.

AddressBuilder

The address builder.

AddressOutput

An Address output.

AddressWrapper

An address and its network type.

Ed25519Address

An Ed25519 address.

UtxoInput

Represents an input referencing an output.

Enums

Input

A generic input supporting different input kinds.

IotaAddress

A generic address supporting different address kinds.

OutputKind

The address output kind.

Functions

get_address_with_index

Gets a public address for the given account.

get_iota_address

get_new_address

Gets an unused public address for the given account.

get_new_change_address

Gets an unused change address for the given account and address.

is_unspent

parse

Parses a bech32 address string.