Struct iota_wallet::client::ClientOptionsBuilder

pub struct ClientOptionsBuilder {
    nodes: Vec<Url>,
    node_pool_urls: Vec<Url>,
    network: Option<String>,
    mqtt_broker_options: Option<BrokerOptions>,
    local_pow: bool,
    node_sync_interval: Option<Duration>,
    node_sync_enabled: bool,
    request_timeout: Option<Duration>,
    api_timeout: HashMap<Api, Duration>,
}

The options builder for a client connected to multiple nodes.

Fields
nodes: Vec<Url>node_pool_urls: Vec<Url>network: Option<String>mqtt_broker_options: Option<BrokerOptions>local_pow: boolnode_sync_interval: Option<Duration>node_sync_enabled: boolrequest_timeout: Option<Duration>api_timeout: HashMap<Api, Duration>
Implementations

impl ClientOptionsBuilder

pub fn new() -> Self

Initialises a new instance of the builder.

pub fn with_nodes(mut self: Self, nodes: &[&str]) -> Result<Self>

ClientOptions connected to a list of nodes.

Examples

use iota_wallet::client::ClientOptionsBuilder;
let client_options = ClientOptionsBuilder::new()
    .with_nodes(&[
        "https://api.lb-0.testnet.chrysalis2.com",
        "https://api.hornet-1.testnet.chrysalis2.com/",
    ])
    .expect("invalid nodes URLs")
    .build();

pub fn with_node(mut self: Self, node: &str) -> Result<Self>

Adds a node to the node list.

pub fn with_node_pool_urls(
mut self: Self,
node_pool_urls: &[&str]
) -> Result<Self>

Get node list from the node_pool_urls

pub fn with_network<N: Into<String>>(mut self: Self, network: N) -> Self

ClientOptions connected to the default Network pool.

Examples

use iota_wallet::client::ClientOptionsBuilder;
let client_options = ClientOptionsBuilder::new().with_network("testnet2").build();

pub fn with_node_sync_interval(
mut self: Self,
node_sync_interval: Duration
) -> Self

Set the node sync interval

pub fn with_node_sync_disabled(mut self: Self) -> Self

Disables the node syncing process. Every node will be considered healthy and ready to use.

pub fn with_mqtt_mqtt_broker_options(
mut self: Self,
options: BrokerOptions
) -> Self

Sets the MQTT broker options.

pub fn with_local_pow(mut self: Self, local: bool) -> Self

Sets whether the PoW should be done locally or remotely.

pub fn with_request_timeout(mut self: Self, timeout: Duration) -> Self

Sets the request timeout.

pub fn with_api_timeout(mut self: Self, api: Api, timeout: Duration) -> Self

Sets the request timeout for a specific API usage.

pub fn build(self) -> Result<ClientOptions>

Builds the options.