Enum iota_wallet::Error

pub enum Error {

    IoError(Error),
    JsonError(Error),
    StrongholdError(Error),
    ClientError(Box<Error>),
    UrlError(ParseError),
    MessageNotFound,
    InvalidMessageIdLength,
    InvalidAddress,
    InvalidAddressLength,
    StorageDoesntExist,
    InsufficientFunds(u64, u64),
    AccountNotEmpty,
    LatestAccountIsEmpty,
    RecordNotFound,
    InvalidRemainderValueAddress,
    Storage(String),
    Panic(String),
    InvalidMessageId,
    InvalidTransactionId,
    AddressBuildRequiredField(AddressBuildRequiredField),
    AccountInitialiseRequiredField(AccountInitialiseRequiredField),
    BeeMessage(Error),
    InvalidBackupFile,
    InvalidBackupDestination,
    MnemonicEncode(String),
    InvalidMnemonic(String),
    StorageExists,
    StorageAdapterNotSet(PathBuf),
    RecordDecrypt(String),
    RecordEncrypt(String),
    StorageIsEncrypted,
    CannotUseIndexIdentifier,
    NoLedgerSignerError,
    LedgerMiscError,
    LedgerDongleLocked,
    LedgerDeniedByUser,
    LedgerDeviceNotFound,
    LedgerEssenceTooLarge,
    LedgerNetMismatch,
    LedgerMnemonicMismatch,
    AccountAliasAlreadyExists,
    DustError(String),
    LeavingDustError(String),
    InvalidOutputKind(String),
    NodesNotSynced(String),
    FailedToGetRemainder,
    TooManyOutputs(usize, usize),
    ConsolidationRequired(usize, usize),
    InputAddressNotFound,
    LegacyClientError(Box<Error>),
    TernaryError,
    InvalidSeed,
    MigrationDataNotFound,
    MigrationBundleNotFound,
    InputNotFound,
    EmptyInputList,
    SpentAddressOnBundle,
    PoisonError,
    TaskJoinError(JoinError),
    StdThreadJoinError,
    SpentOutputNotFound,
    InvalidParticipations,
    ReqwestError(Error),

}

The wallet error type.

Variants

IoError(Error)

Tuple Fields

0: Error

IO error. (storage, backup, restore)

JsonError(Error)

Tuple Fields

0: Error

serde_json error.

StrongholdError(Error)

Tuple Fields

0: Error

stronghold client error.

ClientError(Box<Error>)

Tuple Fields

0: Box<Error>

iota.rs error.

UrlError(ParseError)

Tuple Fields

0: ParseError

url parse error (client options builder).

MessageNotFound

Message not found.

InvalidMessageIdLength

Message id length response invalid.

InvalidAddress

failed to parse address.

InvalidAddressLength

Address length response invalid.

StorageDoesntExist

Tried to backup but storage file doesn’t exist.

InsufficientFunds(u64, u64)

Tuple Fields

0: u64

1: u64

Insufficient funds to send transfer.

AccountNotEmpty

Account isn’t empty (has history or balance) - can’t delete account.

LatestAccountIsEmpty

Latest account is empty (doesn’t have history and balance) - can’t create account.

RecordNotFound

Account not found

InvalidRemainderValueAddress

invalid remainder value target address defined on RemainderValueStrategy. the address must belong to the account.

Storage(String)

Tuple Fields

0: String

Storage access error.

Panic(String)

Tuple Fields

0: String

Panic error.

InvalidMessageId

Invalid message identifier.

InvalidTransactionId

Invalid transaction identifier.

AddressBuildRequiredField(AddressBuildRequiredField)

Tuple Fields

0: AddressBuildRequiredField

Address build error: required field not filled.

AccountInitialiseRequiredField(AccountInitialiseRequiredField)

Tuple Fields

0: AccountInitialiseRequiredField

Account initialisation error: required field not filled.

BeeMessage(Error)

Tuple Fields

0: Error

Error from bee_message crate.

InvalidBackupFile

Path provided to import_accounts isn’t a valid file

InvalidBackupDestination

Backup destination argument is invalid

MnemonicEncode(String)

Tuple Fields

0: String

Mnemonic generation error.

InvalidMnemonic(String)

Tuple Fields

0: String

Invalid mnemonic error

StorageExists

Can’t import accounts because the storage already exist

StorageAdapterNotSet(PathBuf)

Tuple Fields

0: PathBuf

Storage adapter not defined for the given storage path.

RecordDecrypt(String)

Tuple Fields

0: String

error decrypting stored record using provided encryptionKey

RecordEncrypt(String)

Tuple Fields

0: String

error encrypting stored record using provided encryptionKey

StorageIsEncrypted

Can’t use AccountManager API because the storage is encrypted

CannotUseIndexIdentifier

cannot use index to get account - multiple index sequences found (two or more different signer types stored on accounts)

NoLedgerSignerError

No ledger signer error

LedgerMiscError

Ledger transport error

LedgerDongleLocked

Dongle Locked

LedgerDeniedByUser

Denied by User

LedgerDeviceNotFound

Ledger Device not found

LedgerEssenceTooLarge

Ledger Essence Too Large

LedgerNetMismatch

Ledger transport error

LedgerMnemonicMismatch

Wrong ledger seed error

AccountAliasAlreadyExists

Account alias must be unique.

DustError(String)

Tuple Fields

0: String

Dust error, for example no dust allowance ouptut on an address.

LeavingDustError(String)

Tuple Fields

0: String

Leaving dust error, for example sending 2.5 from 3 Mi, would leave 0.5 (dust) behind.

InvalidOutputKind(String)

Tuple Fields

0: String

Invalid output kind.

NodesNotSynced(String)

Tuple Fields

0: String

Node not synced when creating account or updating client options.

FailedToGetRemainder

Failed to get remainder

TooManyOutputs(usize, usize)

Tuple Fields

0: usize

1: usize

Too many outputs

ConsolidationRequired(usize, usize)

Tuple Fields

0: usize

1: usize

Funds are spread over too many outputs

InputAddressNotFound

Provided input address not found

LegacyClientError(Box<Error>)

Tuple Fields

0: Box<Error>

iota 1.0 client error

TernaryError

Ternary error.

InvalidSeed

Invalid legacy seed.

MigrationDataNotFound

Migration data not found.

MigrationBundleNotFound

Migration bundle not found.

InputNotFound

Input not found with given index.

EmptyInputList

Empty input list on migration bundle creation.

SpentAddressOnBundle

Cannot create bundle when the number of inputs is larger than 1 and there’s a spent input. Spent addresses must be the only input in a bundle.

PoisonError

Mutex lock failed.

TaskJoinError(JoinError)

Tuple Fields

0: JoinError

Tokio task join error

StdThreadJoinError

std thread join error

SpentOutputNotFound

Couldn’t get a spent output from a node.

InvalidParticipations

Participation is invalid

ReqwestError(Error)

Tuple Fields

0: Error

reqwest error