Enum iota_wallet::error::Error

pub enum Error {

    IoError(Error),
    JsonError(Error),
    ClientError(Box<Error>),
    InvalidAddress,
    StorageDoesntExist,
    InsufficientFunds(u64, u64),
    LatestAccountIsEmpty,
    AccountNotFound,
    RecordNotFound,
    Storage(String),
    Panic(String),
    Block(Error),
    BlockDtoError(DtoError),
    RestApiError(Error),
    BackupError(&'static str),
    InvalidMnemonic(String),
    InvalidCoinType(u32, u32),
    StorageExists,
    StorageAdapterNotSet(PathBuf),
    RecordDecrypt(String),
    RecordEncrypt(String),
    StorageIsEncrypted,
    AccountAliasAlreadyExists,
    InvalidOutputKind(String),
    MissingParameter(&'static str),
    FailedToGetRemainder,
    TooManyOutputs(usize, u16),
    TooManyInputs(usize, u16),
    ConsolidationRequired(usize, u16),
    AddressNotFoundInAccount(String),
    MintingFailed(String),
    BurningOrMeltingFailed(String),
    NftNotFoundInUnspentOutputs,
    TaskJoinError(JoinError),
    StdThreadJoinError,
    Blake2b256(&'static str),
    CustomInputError(String),
    ClientNotSet,
    TimeNotSynced(u32, u32),

}

The wallet error type.

Variants

IoError(Error)

IO error. (storage, backup, restore)

JsonError(Error)

serde_json error.

ClientError(Box<Error>)

iota.rs error.

InvalidAddress

failed to parse address.

StorageDoesntExist

Tried to backup but storage file doesn’t exist.

InsufficientFunds(u64, u64)

Insufficient funds to send transaction.

LatestAccountIsEmpty

Latest account is empty (doesn’t have history and balance) - can’t create account.

AccountNotFound

Account not found

RecordNotFound

Record not found

Storage(String)

Storage access error.

Panic(String)

Panic error.

Block(Error)

Error from block crate.

BlockDtoError(DtoError)

Block dtos error

RestApiError(Error)

Rest api error

BackupError(&'static str)

Errors during backup creation or restoring

InvalidMnemonic(String)

Invalid mnemonic error

InvalidCoinType(u32, u32)

Invalid coin type, all accounts need to have the same coin type

StorageExists

Can’t import accounts because the storage already exist

StorageAdapterNotSet(PathBuf)

Storage adapter not defined for the given storage path.

RecordDecrypt(String)

error decrypting stored record using provided encryptionKey

RecordEncrypt(String)

error encrypting stored record using provided encryptionKey

StorageIsEncrypted

Can’t use AccountManager API because the storage is encrypted

AccountAliasAlreadyExists

Account alias must be unique.

InvalidOutputKind(String)

Invalid output kind.

MissingParameter(&'static str)

Missing parameter.

FailedToGetRemainder

Failed to get remainder

TooManyOutputs(usize, u16)

Too many outputs

TooManyInputs(usize, u16)

Too many outputs

ConsolidationRequired(usize, u16)

Funds are spread over too many outputs

AddressNotFoundInAccount(String)

Address not found in account

MintingFailed(String)

Minting failed

BurningOrMeltingFailed(String)

Burning or melting failed

NftNotFoundInUnspentOutputs

Nft not found in unspent outputs

TaskJoinError(JoinError)

Tokio task join error

StdThreadJoinError

std thread join error

Blake2b256(&'static str)

Blake2b256 Error

CustomInputError(String)

Custom input error

ClientNotSet

Client not set error

TimeNotSynced(u32, u32)

Local time doesn’t match the time of the latest timestamp