Module iota_wallet::message

The message module.

Structs

IndexationPayload

A payload which holds an index and associated data.

IotaMessage

Represent the object that nodes gossip around the network.

Message

A message definition.

MessageBuilder

MessageId

A message identifier, the BLAKE2b-256 hash of the message bytes. See https://www.blake2.net/ for more information.

MessageMigratedFundsEntry

Migrated funds entry.

MessageMilestonePayload

Message milestone payload.

MessageMilestonePayloadEssence

Milestone payload essence.

MessageReceiptPayload

Receipt payload.

MessageTailTransactionHash

Tail transaction hash.

MessageTransactionPayload

A transaction payload.

MigratedFundsEntry

Describes funds which were migrated from a legacy network.

MilestoneIndex

A wrapper around a u32 that represents a milestone index.

MilestonePayload

A payload which defines the inclusion set of other messages in the Tangle.

MilestonePayloadEssence

Essence of a milestone payload. This is the signed part of a milestone payload.

MilestoneResponse

Milestone data.

Parents

A Message’s Parents are the MessageIds of the messages it directly approves.

ReceiptPayload

Receipt is a listing of migrated funds.

RegularEssence

A transaction regular essence consuming inputs, creating outputs and carrying an optional payload.

SignatureLockedDustAllowanceOutput

A SignatureLockedDustAllowanceOutput functions like a SignatureLockedSingleOutput but as a special property it is used to increase the allowance/amount of dust outputs on a given address.

SignatureLockedSingleOutput

An output type which can be unlocked via a signature. It deposits onto one single address.

TailTransactionHash

Represents a tail transaction hash of a legacy bundle.

TransactionPayload

A transaction to move funds.

TransactionRegularEssence

Regular essence.

TransactionSignatureLockedDustAllowanceOutput

Dust allowance output.

TransactionSignatureLockedSingleOutput

Signature locked single output.

TransactionUtxoInput

UTXO input.

Transfer

A transfer to make a transaction.

TransferBuilder

A transfer to make a transaction.

TransferOutput

Transfer output.

TreasuryInput

TreasuryInput is an input which references a milestone which generated a TreasuryOutput.

TreasuryOutput

TreasuryOutput is an output which holds the treasury of a network.

TreasuryTransactionPayload

TreasuryTransaction represents a transaction which moves funds from the treasury.

UtxoInput

Represents an input referencing an output.

Value

The transaction Value struct.

Enums

Essence

A generic essence that can represent different types defining transaction essences.

Input

A generic input supporting different input kinds.

MessagePayload

The message’s payload.

MessageType

Message type.

Output

A generic output that can represent different types defining the deposit of funds.

Payload

A generic payload that can represent different types defining message payloads.

RemainderValueStrategy

The strategy to use for the remainder value management when sending funds.

TransactionEssence

The transaction essence.

TransactionInput

Transaction input.

TransactionOutput

The transaction output.

UnlockBlock

Defines the mechanism by which a transaction input is authorized to be consumed.

ValueUnit

Possible Value units.

Constants

MILESTONE_CACHE_RANGE

MILESTONE_ISSUE_RATE_SECS

The node issue a milestone every 10 seconds.

MILESTONE_MERKLE_PROOF_LENGTH

Length of a milestone merkle proof.

MILESTONE_PUBLIC_KEY_LENGTH

Length of a milestone public key.

Functions

default_output_kind

is_internal

milestone_cache

transaction_inputs_belonging_to_account

transaction_outputs_belonging_to_account

Type Definitions

MilestoneCache