Struct iota_wallet::message::Message

pub struct Message {
    pub(crate) id: MessageId,
    pub(crate) version: u64,
    pub(crate) parents: Vec<MessageId>,
    pub(crate) payload_length: usize,
    pub(crate) payload: Payload,
    pub(crate) timestamp: DateTime<Utc>,
    pub(crate) nonce: u64,
    pub(crate) confirmed: Option<bool>,
    pub(crate) broadcasted: bool,
    pub(crate) incoming: bool,
    pub(crate) value: u64,
    pub(crate) remainder_value: u64,
}

A message definition.

Fields
id: MessageId

The message identifier.

version: u64

The message version.

parents: Vec<MessageId>

Message ids this message refers to.

payload_length: usize

Length of the payload.

payload: Payload

Message payload.

timestamp: DateTime<Utc>

The transaction timestamp.

nonce: u64

Transaction nonce.

confirmed: Option<bool>

Whether the transaction is confirmed or not.

broadcasted: bool

Whether the transaction is broadcasted or not.

incoming: bool

Whether the message represents an incoming transaction or not.

value: u64

The message's value.

remainder_value: u64

The message's remainder value sum.

Implementations

impl Message

pub fn id(&self) -> &MessageId

The message identifier.

pub fn version(&self) -> &u64

The message version.

pub fn parents(&self) -> &Vec<MessageId>

Message ids this message refers to.

pub fn payload_length(&self) -> &usize

Length of the payload.

pub fn payload(&self) -> &Payload

Message payload.

pub fn timestamp(&self) -> &DateTime<Utc>

The transaction timestamp.

pub fn nonce(&self) -> &u64

Transaction nonce.

pub fn confirmed(&self) -> &Option<bool>

Whether the transaction is confirmed or not.

pub fn broadcasted(&self) -> &bool

Whether the transaction is broadcasted or not.

pub fn incoming(&self) -> &bool

Whether the message represents an incoming transaction or not.

pub fn value(&self) -> &u64

The message's value.

pub fn remainder_value(&self) -> &u64

The message's remainder value sum.

impl Message

pub(crate) fn set_id(&mut self, val: MessageId) -> &mut Self

The message identifier.

pub(crate) fn set_version(&mut self, val: u64) -> &mut Self

The message version.

pub(crate) fn set_parents(&mut self, val: Vec<MessageId>) -> &mut Self

Message ids this message refers to.

pub(crate) fn set_payload_length(&mut self, val: usize) -> &mut Self

Length of the payload.

pub(crate) fn set_payload(&mut self, val: Payload) -> &mut Self

Message payload.

pub(crate) fn set_timestamp(&mut self, val: DateTime<Utc>) -> &mut Self

The transaction timestamp.

pub(crate) fn set_nonce(&mut self, val: u64) -> &mut Self

Transaction nonce.

pub fn set_confirmed(&mut self, val: Option<bool>) -> &mut Self

Whether the transaction is confirmed or not.

pub fn set_broadcasted(&mut self, val: bool) -> &mut Self

Whether the transaction is broadcasted or not.

pub(crate) fn set_incoming(&mut self, val: bool) -> &mut Self

Whether the message represents an incoming transaction or not.

pub(crate) fn set_value(&mut self, val: u64) -> &mut Self

The message's value.

pub(crate) fn set_remainder_value(&mut self, val: u64) -> &mut Self

The message's remainder value sum.

impl Message

pub(crate) fn from_iota_message(
id: MessageId,
account_addresses: &[Address],
message: IotaMessage,
confirmed: Option<bool>
) -> Self

pub fn addresses(&self) -> Vec<&IotaAddress>

The message's addresses.

pub fn compute_value(
iota_message: &IotaMessage,
id: &MessageId,
account_addresses: &[Address]
) -> Value

Gets the absolute value of the transaction.