Trait iota_wallet::message::Packable

pub trait Packable {
    type Error: Debug;
    pub fn packed_len(&self) -> usize;

    pub fn pack<W>(&self, writer: &mut W) -> Result<(), Self::Error>
    where
        W: Write;

    pub fn unpack<R>(reader: &mut R) -> Result<Self, Self::Error>
    where
        R: Read + ?Sized;

    pub fn pack_new(&self) -> Vec<u8, Global> { ... }
}

A trait to pack and unpack types to and from bytes.

Associated Types

type Error: Debug

Associated error type.

Loading content...
Required methods

pub fn packed_len(&self) -> usize

Returns the length of the packed bytes.

pub fn pack<W>(&self, writer: &mut W) -> Result<(), Self::Error> where
W: Write,

Packs the instance to bytes and writes them to the passed writer.

pub fn unpack<R>(reader: &mut R) -> Result<Self, Self::Error> where
R: Read + ?Sized,

Reads bytes from the passed reader and unpacks them into an instance.

Loading content...
Provided methods

pub fn pack_new(&self) -> Vec<u8, Global>

Packs the instance to bytes and writes them to a newly allocated vector.

Loading content...