Module iota_wallet::message_interface

The message passing interface for the library. A different way to call the wallet functions, useful for bindings to other languages.

Modules

account_method

dtos

message

message_handler

message_type

response

Structs

AccountBalanceDto

Dto for the balance of an account, returned from crate::account::handle::AccountHandle::sync() and crate::account::handle::AccountHandle::balance().

AccountToCreate

An account to create.

AddressWithAmountDto

Dto for address with amount for send_amount()

AddressWithUnspentOutputsDto

Dto for an account address with output_ids of unspent outputs.

ManagerOptions

Message

The message type.

WalletMessageHandler

The Wallet message handler.

Enums

AccountMethod

Each public account method.

MessageType

The messages that can be sent to the actor.

Response

The response message.

Functions

create_message_handler

listen

Listen to wallet events, empty vec will listen to all events

secret_manager_serialize

send_message