Enum iota_wallet::message_interface::AccountMethod

pub enum AccountMethod {

    GenerateAddresses {
        amount: u32,
        options: Option<AddressGenerationOptions>,
    },
    GetOutput {
        output_id: OutputId,
    },
    GetOutputsWithAdditionalUnlockConditions {
        outputs_to_collect: OutputsToCollect,
    },
    ListAddresses,
    ListAddressesWithUnspentOutputs,
    ListOutputs,
    ListUnspentOutputs,
    ListTransactions,
    ListPendingTransactions,
    MintNativeToken {
        native_token_options: NativeTokenOptions,
        options: Option<TransferOptions>,
    },
    MintNfts {
        nfts_options: Vec<NftOptions>,
        options: Option<TransferOptions>,
    },
    GetBalance,
    PrepareTransaction {
        outputs: Vec<OutputDto>,
        options: Option<TransferOptions>,
    },
    PrepareMintNfts {
        nfts_options: Vec<NftOptions>,
        options: Option<TransferOptions>,
    },
    PrepareSendAmount {
        addresses_with_amount: Vec<AddressWithAmountDto>,
        options: Option<TransferOptions>,
    },
    PrepareSendMicroTransaction {
        addresses_with_micro_amount: Vec<AddressWithMicroAmountDto>,
        options: Option<TransferOptions>,
    },
    PrepareSendNativeTokens {
        addresses_native_tokens: Vec<AddressNativeTokens>,
        options: Option<TransferOptions>,
    },
    PrepareSendNft {
        addresses_nft_ids: Vec<AddressAndNftId>,
        options: Option<TransferOptions>,
    },
    SyncAccount {
        options: Option<SyncOptions>,
    },
    SendAmount {
        addresses_with_amount: Vec<AddressWithAmountDto>,
        options: Option<TransferOptions>,
    },
    SendMicroTransaction {
        addresses_with_micro_amount: Vec<AddressWithMicroAmountDto>,
        options: Option<TransferOptions>,
    },
    SendNativeTokens {
        addresses_native_tokens: Vec<AddressNativeTokens>,
        options: Option<TransferOptions>,
    },
    SendNft {
        addresses_nft_ids: Vec<AddressAndNftId>,
        options: Option<TransferOptions>,
    },
    SetAlias {
        alias: String,
    },
    SendTransfer {
        outputs: Vec<OutputDto>,
        options: Option<TransferOptions>,
    },
    SignTransactionEssence {
        prepared_transaction_data: PreparedTransactionDataDto,
    },
    SubmitAndStoreTransaction {
        signed_transaction_data: SignedTransactionDataDto,
    },
    TryCollectOutputs {
        outputs_to_collect: OutputsToCollect,
    },
    CollectOutputs {
        output_ids_to_collect: Vec<OutputId>,
    },

}

Each public account method.

Variants

GenerateAddresses

Fields

amount: u32

options: Option<AddressGenerationOptions>

Generate a new unused address. Expected response: GeneratedAddress

GetOutput

Fields

output_id: OutputId

Get the OutputData of an output stored in the account Expected response: Output

GetOutputsWithAdditionalUnlockConditions

Fields

outputs_to_collect: OutputsToCollect

Get outputs with additional unlock conditions Expected response: OutputIds

ListAddresses

Expected response: Addresses List addresses.

ListAddressesWithUnspentOutputs

Returns only addresses of the account with unspent outputs Expected response: AddressesWithUnspentOutputs

ListOutputs

Returns all outputs of the account Expected response: Outputs

ListUnspentOutputs

Returns all unspent outputs of the account Expected response: Outputs

ListTransactions

Returns all transaction of the account Expected response: Transactions

ListPendingTransactions

Returns all pending transaction of the account Expected response: Transactions

MintNativeToken

Fields

native_token_options: NativeTokenOptions

options: Option<TransferOptions>

Mint native token. Expected response: SentTransfer

MintNfts

Fields

nfts_options: Vec<NftOptions>

options: Option<TransferOptions>

Mint nft. Expected response: SentTransfer

GetBalance

Get account balance information. Expected response: Balance

PrepareTransaction

Fields

outputs: Vec<OutputDto>

options: Option<TransferOptions>

Prepare transaction. Expected response: PreparedTransactionData

PrepareMintNfts

Fields

nfts_options: Vec<NftOptions>

options: Option<TransferOptions>

Prepare mint nft. Expected response: PreparedTransactionData

PrepareSendAmount

Fields

addresses_with_amount: Vec<AddressWithAmountDto>

options: Option<TransferOptions>

Prepare send amount. Expected response: PreparedTransactionData

PrepareSendMicroTransaction

Fields

addresses_with_micro_amount: Vec<AddressWithMicroAmountDto>

options: Option<TransferOptions>

Prepare send amount below minimum storage deposit. Expected response: PreparedTransactionData

PrepareSendNativeTokens

Fields

addresses_native_tokens: Vec<AddressNativeTokens>

options: Option<TransferOptions>

Prepare send native tokens. Expected response: PreparedTransactionData

PrepareSendNft

Fields

addresses_nft_ids: Vec<AddressAndNftId>

options: Option<TransferOptions>

Prepare send nft. Expected response: PreparedTransactionData

SyncAccount

Fields

options: Option<SyncOptions>

Sync options

Syncs the account by fetching new information from the nodes. Will also retry pending transactions and consolidate outputs if necessary. Expected response: Balance

SendAmount

Fields

addresses_with_amount: Vec<AddressWithAmountDto>

options: Option<TransferOptions>

Send amount. Expected response: SentTransfer

SendMicroTransaction

Fields

addresses_with_micro_amount: Vec<AddressWithMicroAmountDto>

options: Option<TransferOptions>

Send amount below minimum storage deposit. Expected response: SentTransfer

SendNativeTokens

Fields

addresses_native_tokens: Vec<AddressNativeTokens>

options: Option<TransferOptions>

Send native tokens. Expected response: SentTransfer

SendNft

Fields

addresses_nft_ids: Vec<AddressAndNftId>

options: Option<TransferOptions>

Send nft. Expected response: SentTransfer

SetAlias

Fields

alias: String

Set the alias of the account. Expected response: Ok

SendTransfer

Fields

outputs: Vec<OutputDto>

options: Option<TransferOptions>

Send funds. Expected response: SentTransfer

SignTransactionEssence

Fields

prepared_transaction_data: PreparedTransactionDataDto

Sign a prepared transaction. Expected response: TransactionPayload

SubmitAndStoreTransaction

Fields

signed_transaction_data: SignedTransactionDataDto

Validate the transaction, submit it to a node and store it in the account. Expected response: SentTransfer

TryCollectOutputs

Fields

outputs_to_collect: OutputsToCollect

Try to collect outputs. Expected response: SentTransfers

CollectOutputs

Fields

output_ids_to_collect: Vec<OutputId>

Collect outputs. Expected response: SentTransfers