Module iota_wallet::secret

Secret manager module enabling address generation and transaction essence signing.

Modules

ledger_nano

Implementation of LedgerSecretManager.

mnemonic

Module for signing with a mnemonic or seed Implementation of MnemonicSecretManager.

placeholder

Module for the PlaceholderSecretManager Implementation of PlaceholderSecretManager.

stronghold

Module for signing with a Stronghold vault Stronghold-as-a-Secret-Manager.

types

Signing related types Miscellaneous types for secret managers.

Structs

GenerateAddressMetadata

Metadata provided to generate_address().

LedgerStatus

The Ledger device status.

Enums

SecretManager

Supported secret managers

SecretManagerDto

DTO for secret manager types with required data.

Traits

SecretManage

The secret manager interface.

SecretManageExt

An extension to SecretManager.