Trait iota_wallet::secret::SecretManageExt

pub trait SecretManageExt {
    fn sign_transaction_essence<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        prepared_transaction_data: &'life1 PreparedTransactionData
    ) -> Pin<Box<dyn Future<Output = Result<Unlocks, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

An extension to SecretManager.

This trait is automatically implemented for any type that implements SecretManager - it contains methods for internal use that are based on the methods in SecretManager. Secret managers don’t implement this on their sides.

Required Methods

fn sign_transaction_essence<'life0, 'life1, 'async_trait>(
&'life0 self,
prepared_transaction_data: &'life1 PreparedTransactionData
) -> Pin<Box<dyn Future<Output = Result<Unlocks, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Signs transaction essence.

Secret managers usually don’t implement this, as the default implementation has taken care of the placement of blocks (e.g. references between them). SecretManager::signature_unlock() will be invoked every time a necessary signing action needs to be performed.