Module iota_wallet::stronghold

Stronghold interface abstractions over an account

Structs

ActorRuntime

Status
Stronghold status.

Enums

Error

SnapshotStatus
Snapshot status.

Constants

ACCOUNT_IDS_RECORD_PATH

ACCOUNT_ID_SEPARATOR

ACCOUNT_METADATA_VAULT_PATH

ACCOUNT_RECORD_PATH

ACCOUNT_VAULT_PATH

DEFAULT_PASSWORD_CLEAR_INTERVAL

DERIVE_OUTPUT_RECORD_PATH

SECRET_VAULT_PATH

SEED_RECORD_PATH

Statics

CURRENT_SNAPSHOT_PATH

PASSWORD_CLEAR_INTERVAL

PASSWORD_STORE

PRIVATE_DATA_CLIENT_PATH

STRONGHOLD_ACCESS_STORE

Functions

account_id_to_client_path

actor_runtime

change_password
Changes the snapshot password.

check_snapshot

clear_stronghold_cache

default_password_store

derive

generate_address

get_account

get_account_internal

get_accounts

get_password

get_public_key

get_status
Gets the stronghold status for the given snapshot.

load_account_actor

load_actor

load_actors

load_private_data_actor

load_snapshot

load_snapshot_internal

on_stronghold_access

remove_account

save_snapshot

set_password

set_password_clear_interval
Set the password clear interval. If the stronghold isn't used after interval, the password is cleared and must be set again.

sign_essence

store_account

store_mnemonic

stronghold_response_to_result

switch_snapshot

unload_snapshot
Removes the snapshot from memory and clears the password.

Type Definitions

Result

SnapshotToPasswordMap