Struct iota_wallet::Local

pub struct Local;

The local timescale. This is implemented via the standard time crate.

Using the TimeZone methods on the Local struct is the preferred way to construct DateTime<Local> instances.

Example

use chrono::{Local, DateTime, TimeZone};

let dt: DateTime<Local> = Local::now();
let dt: DateTime<Local> = Local.timestamp(0, 0);

Implementations

impl Local

pub fn today() -> Date<Local>

Returns a Date which corresponds to the current date.

pub fn now() -> DateTime<Local>

Returns a DateTime which corresponds to the current date.