Crate iota_wallet

The IOTA Wallet Library

Modules

account
The account module.

account_manager
The account manager module.

actor
The actor interface for the library.

address
The address module.

client
The client module.

error

event
The event module.

message
The message module.

monitor
The monitor module.

serde

signing
Signing interfaces.

storage
The storage module.

stronghold
Stronghold interface abstractions over an account

Structs

DateTime
ISO 8601 combined date and time with time zone.

Local
The local timescale. This is implemented via the standard time crate.

StrongholdStatus
Stronghold status.

Utc
The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Enums

Error
The wallet error type.

LedgerStatus
The Ledger device status.

StrongholdSnapshotStatus
Snapshot status.

Statics

RUNTIME

Functions

block_on

get_ledger_status
Gets the status of the Ledger device/simulator.

get_stronghold_status
Gets the stronghold status for the given snapshot.

lock_stronghold
Removes the snapshot from memory and clears the password.

set_stronghold_password_clear_interval
Set the password clear interval. If the stronghold isn’t used after interval, the password is cleared and must be set again.

spawn

with_actor_system
Access the stronghold’s actor system.

Type Definitions

Result
The wallet Result type.