Module iota_wallet::account

The account module. Interaction with an Account happens via an AccountHandle.

Modules

builder 🔒

The module with the AccountBuilder.

constants 🔒

Constants used for the account and account operations.

handle 🔒

A thread guard over an account, all account methods are called from here.

operations 🔒

The account operations like address generation, syncing and creating transfers.

types

Types used in an account and returned from methods.

Structs

Account

An Account.

AccountHandle

A thread guard over an account, so we can lock the account during operations.

AccountOptions

Account options

AddressGenerationOptions

Options for address generation

SyncOptions

The synchronization options

TransferOptions

Options for value transfers

Enums

OutputsToCollect

Enum to specify which outputs should be collected

RemainderValueStrategy

The strategy to use for the remainder value management when sending funds.