Struct iota_wallet::account::AccountHandle

pub struct AccountHandle {
    inner: Arc<RwLock<Account>>,
    locked_addresses: Arc<Mutex<Vec<AddressWrapper>>>,
}

A thread guard over an account.

Fields
inner: Arc<RwLock<Account>>locked_addresses: Arc<Mutex<Vec<AddressWrapper>>>
Implementations

impl AccountHandle

pub(crate) fn locked_addresses(&self) -> Arc<Mutex<Vec<AddressWrapper>>>

impl AccountHandle

pub async fn id(&self) -> String

Bridge to Account#id.

pub async fn signer_type(&self) -> SignerType

Bridge to Account#signer_type.

pub async fn index(&self) -> usize

Bridge to Account#index.

pub async fn alias(&self) -> String

Bridge to Account#alias.

pub async fn created_at(&self) -> DateTime<Local>

Bridge to Account#created_at.

pub async fn last_synced_at(&self) -> Option<DateTime<Local>>

Bridge to Account#last_synced_at.

pub async fn messages(&self) -> Vec<Message>

Bridge to Account#messages. This method clones the messages so prefer the using the read method to access the account instance.

pub async fn addresses(&self) -> Vec<Address>

Bridge to Account#addresses. This method clones the addresses so prefer the using the read method to access the account instance.

pub async fn client_options(&self) -> ClientOptions

Bridge to Account#client_options.

impl AccountHandle

pub async fn sync(&self) -> AccountSynchronizer

Returns the builder to setup the process to synchronize this account with the Tangle.

pub async fn generate_address(&self) -> Result<Address>

Gets a new unused address and links it to this account.

pub(crate) async fn generate_address_internal(
&self,
account: &mut RwLockWriteGuard<'_, Account>
) -> Result<Address>

Generates an address without locking the account.

pub async fn get_unused_address(&self) -> Result<Address>

Synchronizes the account addresses with the Tangle and returns the latest address in the account, which is an address without balance.

pub async fn is_latest_address_unused(&self) -> Result<bool>

Syncs the latest address with the Tangle and determines whether it's unused or not. An unused address is an address without balance and associated message history. Note that such address might have been used in the past, because the message history might have been pruned by the node.

pub async fn latest_address(&self) -> Address

Bridge to Account#latest_address.

pub async fn balance(&self) -> AccountBalance

Bridge to Account#balance.

pub async fn set_alias(&self, alias: impl AsRef<str>) -> Result<()>

Bridge to Account#set_alias.

pub async fn set_client_options(&self, options: ClientOptions) -> Result<()>

Bridge to Account#set_client_options.

pub async fn list_messages(
&self,
count: usize,
from: usize,
message_type: Option<MessageType>
) -> Vec<Message>

Bridge to Account#list_messages. This method clones the account's messages so when querying a large list of messages prefer using the read method to access the account instance.

pub async fn list_spent_addresses(&self) -> Vec<Address>

Bridge to Account#list_spent_addresses. This method clones the account's addresses so when querying a large list of addresses prefer using the read method to access the account instance.

pub async fn list_unspent_addresses(&self) -> Vec<Address>

Bridge to Account#list_unspent_addresses. This method clones the account's addresses so when querying a large list of addresses prefer using the read method to access the account instance.

pub async fn get_message(&self, message_id: &MessageId) -> Option<Message>

Bridge to Account#get_message.