Struct iota_wallet::account_manager::AccountManager

pub struct AccountManager {
    storage_folder: PathBuf,
    loaded_accounts: AtomicBool,
    storage_path: PathBuf,
    accounts: AccountStore,
    stop_polling_sender: StdMutex<Option<BroadcastSender<()>>>,
    polling_handle: StdMutex<Option<JoinHandle<()>>>,
    generated_mnemonic: StdMutex<Option<String>>,
    account_options: AccountOptions,
    sync_accounts_lock: Arc<Mutex<()>>,
    cached_migration_data: Mutex<HashMap<u64, CachedMigrationData>>,
    cached_migration_bundles: Mutex<HashMap<String, (Vec<BundledTransaction>, AddressWrapper, u64)>>,
}

The account manager.

Used to manage multiple accounts.

Fields
storage_folder: PathBufloaded_accounts: AtomicBoolstorage_path: PathBuf

the path to the storage.

accounts: AccountStore

Returns a handle to the accounts store.

stop_polling_sender: StdMutex<Option<BroadcastSender<()>>>polling_handle: StdMutex<Option<JoinHandle<()>>>generated_mnemonic: StdMutex<Option<String>>account_options: AccountOptionssync_accounts_lock: Arc<Mutex<()>>cached_migration_data: Mutex<HashMap<u64, CachedMigrationData>>cached_migration_bundles: Mutex<HashMap<String, (Vec<BundledTransaction>, AddressWrapper, u64)>>
Implementations

impl AccountManager

pub fn storage_path(&self) -> &PathBuf

the path to the storage.

pub fn accounts(&self) -> &AccountStore

Returns a handle to the accounts store.

impl AccountManager

pub fn builder() -> AccountManagerBuilder

Initialises the account manager builder.

pub async fn get_migration_data(
&self,
finder: MigrationDataFinder<'_>
) -> Result<MigrationData>

Gets the legacy migration data for the seed.

pub async fn get_ledger_migration_data(
&self,
addresses: Vec<AddressInput>,
nodes: Vec<&str>,
permanode: Option<String>
) -> Result<MigrationData>

Gets ledger legacy migration data for provided addresses.

pub async fn get_migration_address<I: Into<AccountIdentifier>>(
&self,
ledger_prompt: bool,
account_id: I
) -> Result<MigrationAddress>

Convert the first address from the first account to a migration tryte address

pub async fn mine_bundle(
&self,
prepared_bundle: Vec<String>,
spent_bundle_hashes: Vec<String>,
security_level: u8,
timeout: u64,
offset: i64
) -> Result<MinedBundle>

Mine bundle

pub async fn create_migration_bundle(
&self,
seed: &str,
input_address_indexes: &[u64],
mine: bool,
timeout: Duration,
offset: i64,
log_file_name: &str
) -> Result<MigrationBundle>

Creates the bundle for migration associated with the given input indexes, Performs bundle mining if the address is spent and mine is true, And signs the bundle. Returns the bundle hash. It logs the operations to $storage_path.join(log_file_name).

pub async fn send_migration_bundle(
&self,
nodes: &[&str],
hash: &str,
mwm: u8
) -> Result<MigratedBundle>

Sends the migration bundle to the given node.

pub async fn send_ledger_migration_bundle(
&self,
nodes: &[&str],
bundle: Vec<String>,
mwm: u8
) -> Result<MigratedBundle>

Sends the migration bundle to the given node.

async fn unload_accounts(accounts: &AccountStore) -> Result<()>

async fn load_accounts(
accounts: &AccountStore,
storage_file_path: &Path,
account_options: AccountOptions,
sync_accounts_lock: Arc<Mutex<()>>
) -> Result<()>

pub async fn delete(self) -> Result<(), (Error, Self)>

Deletes the storage.

pub(crate) async fn delete_internal(&self) -> Result<()>

pub(crate) async fn stronghold_snapshot_path(&self) -> Result<PathBuf>

pub(crate) async fn stronghold_snapshot_path_internal(
&self,
storage_id: &str
) -> Result<PathBuf>

fn check_storage_encryption(&self) -> Result<()>

async fn start_monitoring(accounts: AccountStore)

Starts monitoring the accounts with the node’s mqtt topics.

pub async fn start_background_sync(
&self,
polling_interval: Duration,
automatic_output_consolidation: bool
) -> Result<()>

Initialises the background polling and MQTT monitoring.

pub fn stop_background_sync(&self) -> Result<()>

Stops the background polling and MQTT monitoring.

pub async fn clear_storage_password(&self) -> Result<()>

Clear the encryption key and then unload decrypted accounts in memory. Does nothing if storage is not encrypted

pub async fn set_storage_password<P: AsRef<str>>(
&self,
password: P
) -> Result<()>

Sets the password for the stored accounts.

pub async fn set_stronghold_password<P: Into<String>>(
&self,
password: P
) -> Result<()>

Sets the stronghold password.

pub async fn change_stronghold_password<C: Into<String>, N: Into<String>>(
&self,
current_password: C,
new_password: N
) -> Result<()>

Changes the stronghold password.

pub async fn is_latest_address_unused(&self) -> Result<bool>

Determines whether all accounts has the latest address unused.

pub async fn set_client_options(&self, options: ClientOptions) -> Result<()>

Sets the client options for all accounts.

fn start_polling(
&self,
polling_interval: Duration,
stop: BroadcastReceiver<()>,
automatic_output_consolidation: bool
) -> Result<()>

Starts the polling mechanism.

pub async fn store_mnemonic(
&self,
signer_type: SignerType,
mnemonic: Option<String>
) -> Result<()>

Stores a mnemonic for the given signer type. If the mnemonic is not provided, we’ll generate one.

pub fn generate_mnemonic(&self) -> Result<String>

Generates a new mnemonic.

pub fn verify_mnemonic<S: AsRef<str>>(&self, mnemonic: S) -> Result<()>

Checks is the mnemonic is valid. If a mnemonic was generated with generate_mnemonic(), the mnemonic here should match the generated.

pub fn create_account(
&self,
client_options: ClientOptions
) -> Result<AccountInitialiser>

Adds a new account.

pub async fn remove_account<I: Into<AccountIdentifier>>(
&self,
account_id: I
) -> Result<()>

Deletes an account.

pub fn sync_accounts(&self) -> Result<AccountsSynchronizer>

Syncs all accounts.

pub async fn internal_transfer<F: Into<AccountIdentifier>, T: Into<AccountIdentifier>>(
&self,
from_account_id: F,
to_account_id: T,
amount: NonZeroU64
) -> Result<Message>

Transfers an amount from an account to another.

pub async fn backup<P: AsRef<Path>>(
&self,
destination: P,
stronghold_password: String
) -> Result<PathBuf>

Backups the storage to the given destination

pub async fn import_accounts<S: AsRef<Path>>(
&self,
source: S,
stronghold_password: String
) -> Result<()>

Import backed up accounts.

pub async fn get_account<I: Into<AccountIdentifier>>(
&self,
account_id: I
) -> Result<AccountHandle>

Gets the account associated with the given identifier.

pub async fn get_accounts(&self) -> Result<Vec<AccountHandle>>

Gets all accounts from storage.

pub async fn reattach<I: Into<AccountIdentifier>>(
&self,
account_id: I,
message_id: &MessageId
) -> Result<Message>

Reattaches an unconfirmed transaction.

pub async fn promote<I: Into<AccountIdentifier>>(
&self,
account_id: I,
message_id: &MessageId
) -> Result<Message>

Promotes an unconfirmed transaction.

pub async fn retry<I: Into<AccountIdentifier>>(
&self,
account_id: I,
message_id: &MessageId
) -> Result<Message>

Retries an unconfirmed transaction.

pub fn get_seed_checksum(seed: String) -> Result<String>

Get seed checksum

pub async fn participate(
&self,
account_identifier: AccountIdentifier,
participations: Vec<Participation>
) -> Result<Vec<Message>>

Participate in events

pub async fn stop_participating(
&self,
account_identifier: AccountIdentifier,
event_ids: Vec<String>
) -> Result<Vec<Message>>

Stop participating from provided events

pub async fn get_participation_overview(&self) -> Result<ParticipatingAccounts>

Get a participating overview of the accounts by looking at the messages

pub async fn get_participation_events(&self) -> Result<Vec<EventData>>

Get a participating events data

impl AccountManager

pub async fn get_balance_change_events<T: Into<Option<i64>>>(
&self,
count: usize,
skip: usize,
from_timestamp: T
) -> Result<Vec<BalanceEvent>>

Gets the paginated events with an optional timestamp filter.

pub async fn get_balance_change_event_count<T: Into<Option<i64>>>(
&self,
from_timestamp: T
) -> Result<usize>

Gets the count of events with an optional timestamp filter.

impl AccountManager

pub async fn get_transaction_confirmation_events<T: Into<Option<i64>>>(
&self,
count: usize,
skip: usize,
from_timestamp: T
) -> Result<Vec<TransactionConfirmationChangeEvent>>

Gets the paginated events with an optional timestamp filter.

pub async fn get_transaction_confirmation_event_count<T: Into<Option<i64>>>(
&self,
from_timestamp: T
) -> Result<usize>

Gets the count of events with an optional timestamp filter.

impl AccountManager

pub async fn get_new_transaction_events<T: Into<Option<i64>>>(
&self,
count: usize,
skip: usize,
from_timestamp: T
) -> Result<Vec<TransactionEvent>>

Gets the paginated events with an optional timestamp filter.

pub async fn get_new_transaction_event_count<T: Into<Option<i64>>>(
&self,
from_timestamp: T
) -> Result<usize>

Gets the count of events with an optional timestamp filter.

impl AccountManager

pub async fn get_reattachment_events<T: Into<Option<i64>>>(
&self,
count: usize,
skip: usize,
from_timestamp: T
) -> Result<Vec<TransactionReattachmentEvent>>

Gets the paginated events with an optional timestamp filter.

pub async fn get_reattachment_event_count<T: Into<Option<i64>>>(
&self,
from_timestamp: T
) -> Result<usize>

Gets the count of events with an optional timestamp filter.

impl AccountManager

pub async fn get_broadcast_events<T: Into<Option<i64>>>(
&self,
count: usize,
skip: usize,
from_timestamp: T
) -> Result<Vec<TransactionEvent>>

Gets the paginated events with an optional timestamp filter.

pub async fn get_broadcast_event_count<T: Into<Option<i64>>>(
&self,
from_timestamp: T
) -> Result<usize>

Gets the count of events with an optional timestamp filter.