Enum iota_wallet::actor::message::MessageType

pub enum MessageType {

    RemoveAccount(AccountIdentifier),
    CreateAccount(Box<AccountToCreate>),
    GetAccount(AccountIdentifier),
    GetAccounts,
    CallAccountMethod {
        account_id: AccountIdentifier,
        method: AccountMethod,
    },
    SyncAccounts {
        address_index: Option<usize>,
        gap_limit: Option<usize>,
        account_discovery_threshold: Option<usize>,
    },
    Reattach {
        account_id: AccountIdentifier,
        message_id: String,
    },
    Backup {
        destination: PathBuf,
        password: String,
    },
    RestoreBackup {
        backup_path: String,
        password: String,
    },
    SetStoragePassword(String),
    SetStrongholdPassword(String),
    SetStrongholdPasswordClearInterval(Duration),
    GetStrongholdStatus,
    LockStronghold,
    SendTransfer {
        account_id: AccountIdentifier,
        transfer: Box<TransferBuilder>,
    },
    InternalTransfer {
        from_account_id: AccountIdentifier,
        to_account_id: AccountIdentifier,
        amount: NonZeroU64,
    },
    GenerateMnemonic,
    VerifyMnemonic(String),
    StoreMnemonic {
        signer_type: SignerType,
        mnemonic: Option<String>,
    },
    IsLatestAddressUnused,
    GetLedgerStatus(bool),
    DeleteStorage,
    ChangeStrongholdPassword {
        current_password: String,
        new_password: String,
    },
    SetClientOptions(Box<ClientOptions>),
    GetMigrationData {
        nodes: Vec<String>,
        permanode: Option<String>,
        seed: String,
        security_level: Option<u8>,
        initial_address_index: Option<u64>,
    },
    CreateMigrationBundle {
        seed: String,
        input_address_indexes: Vec<u64>,
        mine: bool,
        timeout_secs: u64,
        offset: i64,
        log_file_name: String,
    },
    SendMigrationBundle {
        nodes: Vec<String>,
        bundle_hash: String,
        mwm: u8,
    },
    GetSeedChecksum(String),

}

The messages that can be sent to the actor.

Variants

RemoveAccount(AccountIdentifier)

Remove the account related to the specified account_id.

CreateAccount(Box<AccountToCreate>)

Creates an account.

GetAccount(AccountIdentifier)

Read account.

GetAccounts

Read accounts.

CallAccountMethod

Consume an account method.

Fields of CallAccountMethod

account_id: AccountIdentifier

The account identifier.

method: AccountMethod

The account method to call.

SyncAccounts

Sync accounts.

Fields of SyncAccounts

address_index: Option<usize>

The first address index to sync.

gap_limit: Option<usize>

The gap limit.

account_discovery_threshold: Option<usize>

Minimum number of accounts to check on discovery.

Reattach

Reattach message.

Fields of Reattach

account_id: AccountIdentifier

The account identifier.

message_id: String

The message to reattach.

Backup

Backup storage.

Fields of Backup

destination: PathBuf

The backup destination.

password: String

Stronghold file password.

RestoreBackup

Import accounts from storage.

Fields of RestoreBackup

backup_path: String

The path to the backed up storage.

password: String

Stronghold file password.

SetStoragePassword(String)

Sets the password used to encrypt/decrypt the storage.

SetStrongholdPassword(String)

Set stronghold snapshot password.

SetStrongholdPasswordClearInterval(Duration)

Sets the password clear interval.

GetStrongholdStatus

Get stronghold status.

LockStronghold

Lock the stronghold snapshot (clears password and unload snapshot from memory).

SendTransfer

Send funds.

Fields of SendTransfer

account_id: AccountIdentifier

The account identifier.

transfer: Box<TransferBuilder>

The transfer details.

InternalTransfer

Move funds on stored accounts.

Fields of InternalTransfer

from_account_id: AccountIdentifier

The source account identifier.

to_account_id: AccountIdentifier

The destination account identifier.

amount: NonZeroU64

The transfer amount.

GenerateMnemonic

Generates a new mnemonic.

VerifyMnemonic(String)

Checks if the given mnemonic is valid.

StoreMnemonic

Store mnemonic.

Fields of StoreMnemonic

signer_type: SignerType

The signer type.

mnemonic: Option<String>

The mnemonic. If empty, we’ll generate one.

IsLatestAddressUnused

Checks if all accounts has unused latest address after syncing with the Tangle.

GetLedgerStatus(bool)

Get the Ledger Nano or Speculos simulator status.

DeleteStorage

Deletes the storage.

ChangeStrongholdPassword

Changes stronghold snapshot password.

Fields of ChangeStrongholdPassword

current_password: String

The current stronghold password.

new_password: String

The new stronghold password.

SetClientOptions(Box<ClientOptions>)

Updates the client options for all accounts.

GetMigrationData

Get legacy network balance for the seed.

Fields of GetMigrationData

nodes: Vec<String>

The nodes to connect to.

permanode: Option<String>

The permanode to use.

seed: String

The legacy seed.

security_level: Option<u8>

The WOTS address security level.

initial_address_index: Option<u64>

The initial address index.

CreateMigrationBundle

Creates the bundle for migration, performs bundle mining if the address was spent and signs the bundle.

Fields of CreateMigrationBundle

seed: String

The legacy seed.

input_address_indexes: Vec<u64>

The bundle input address indexes.

mine: bool

Whether we should perform bundle mining or not.

timeout_secs: u64

Timeout in seconds for the bundle mining process.

offset: i64

Offset for the bundle mining process.

log_file_name: String

The name of the log file (stored on the storage folder).

SendMigrationBundle

Sends the migration bundle associated with the hash.

Fields of SendMigrationBundle

nodes: Vec<String>

Node URLs.

bundle_hash: String

Bundle hash returned on CreateMigrationBundle.

mwm: u8

Minimum weight magnitude.

GetSeedChecksum(String)

Get seed checksum.