Enum iota_wallet::actor::message::MessageType

pub enum MessageType {
    RemoveAccount(AccountIdentifier),
    CreateAccount(Box<AccountToCreate>),
    GetAccount(AccountIdentifier),
    GetAccounts,
    CallAccountMethod {
        account_id: AccountIdentifier,
        method: AccountMethod,
    },
    SyncAccounts,
    Reattach {
        account_id: AccountIdentifier,
        message_id: String,
    },
    Backup(String),
    RestoreBackup {
        backup_path: String,
        password: String,
    },
    SetStoragePassword(String),
    SetStrongholdPassword(String),
    SetStrongholdPasswordClearInterval(Duration),
    GetStrongholdStatus,
    LockStronghold,
    SendTransfer {
        account_id: AccountIdentifier,
        transfer: TransferBuilder,
    },
    InternalTransfer {
        from_account_id: AccountIdentifier,
        to_account_id: AccountIdentifier,
        amount: NonZeroU64,
    },
    GenerateMnemonic,
    VerifyMnemonic(String),
    StoreMnemonic {
        signer_type: SignerType,
        mnemonic: Option<String>,
    },
    IsLatestAddressUnused,
    OpenLedgerApp(bool),
    DeleteStorage,
    ChangeStrongholdPassword {
        current_password: String,
        new_password: String,
    },
}

The messages that can be sent to the actor.

Variants

RemoveAccount(AccountIdentifier)

Remove the account related to the specified account_id.

CreateAccount(Box<AccountToCreate>)

Creates an account.

GetAccount(AccountIdentifier)

Read account.

GetAccounts

Read accounts.

CallAccountMethod

Consume an account method.

Fields of CallAccountMethod

account_id: AccountIdentifier

The account identifier.

method: AccountMethod

The account method to call.

SyncAccounts

Sync accounts.

Reattach

Reattach message.

Fields of Reattach

account_id: AccountIdentifier

The account identifier.

message_id: String

The message to reattach.

Backup(String)

Backup storage.

RestoreBackup

Import accounts from storage.

Fields of RestoreBackup

backup_path: String

The path to the backed up storage.

password: String

The backup stronghold password.

SetStoragePassword(String)

Sets the password used to encrypt/decrypt the storage.

SetStrongholdPassword(String)

Set stronghold snapshot password.

SetStrongholdPasswordClearInterval(Duration)

Sets the password clear interval.

GetStrongholdStatus

Get stronghold status.

LockStronghold

Lock the stronghold snapshot (clears password and unload snapshot from memory).

SendTransfer

Send funds.

Fields of SendTransfer

account_id: AccountIdentifier

The account identifier.

transfer: TransferBuilder

The transfer details.

InternalTransfer

Move funds on stored accounts.

Fields of InternalTransfer

from_account_id: AccountIdentifier

The source account identifier.

to_account_id: AccountIdentifier

The destination account identifier.

amount: NonZeroU64

The transfer amount.

GenerateMnemonic

Generates a new mnemonic.

VerifyMnemonic(String)

Checks if the given mnemonic is valid.

StoreMnemonic

Store mnemonic.

Fields of StoreMnemonic

signer_type: SignerType

The signer type.

mnemonic: Option<String>

The mnemonic. If empty, we'll generate one.

IsLatestAddressUnused

Checks if all accounts has unused latest address after syncing with the Tangle.

OpenLedgerApp(bool)

Open the iota ledger app on Ledger Nano or Speculos simulator.

DeleteStorage

Deletes the storage.

ChangeStrongholdPassword

Changes stronghold snapshot password.

Fields of ChangeStrongholdPassword

current_password: String

The current stronghold password.

new_password: String

The new stronghold password.