Enum iota_wallet::error::Error

pub enum Error {
    IoError(Error),
    JsonError(Error),
    StrongholdError(Error),
    ClientError(Error),
    UrlError(ParseError),
    MessageNotFound,
    InvalidMessageIdLength,
    InvalidAddress,
    InvalidAddressLength,
    StorageDoesntExist,
    InsufficientFunds,
    AccountNotEmpty,
    LatestAccountIsEmpty,
    AccountNotFound,
    InvalidRemainderValueAddress,
    Storage(String),
    Panic(String),
    InvalidMessageId,
    InvalidTransactionId,
    AddressBuildRequiredField(AddressBuildRequiredField),
    AccountInitialiseRequiredField(AccountInitialiseRequiredField),
    BeeMessage(Error),
    InvalidBackupFile,
    InvalidBackupDestination,
    StorageAdapterNotDefined,
    MnemonicEncode(String),
    InvalidMnemonic(String),
    StorageExists,
    StorageAdapterNotSet(PathBuf),
    AccountDecrypt(String),
    AccountEncrypt(String),
    StorageIsEncrypted,
    CannotUseIndexIdentifier,
    LedgerMiscError,
    LedgerDongleLocked,
    LedgerDeniedByUser,
    LedgerDeviceNotFound,
    LedgerEssenceTooLarge,
    AccountAliasAlreadyExists,
    DustError(String),
}

The wallet error type.

Variants

IoError(Error)

IO error. (storage, backup, restore)

JsonError(Error)

serde_json error.

StrongholdError(Error)

stronghold client error.

ClientError(Error)

iota.rs error.

UrlError(ParseError)

url parse error (client options builder).

MessageNotFound

Message not found.

InvalidMessageIdLength

Message id length response invalid.

InvalidAddress

failed to parse address.

InvalidAddressLength

Address length response invalid.

StorageDoesntExist

Tried to backup but storage file doesn't exist.

InsufficientFunds

Insufficient funds to send transfer.

AccountNotEmpty

Account isn't empty (has history or balance) - can't delete account.

LatestAccountIsEmpty

Latest account is empty (doesn't have history and balance) - can't create account.

AccountNotFound

Account not found

InvalidRemainderValueAddress

invalid remainder value target address defined on RemainderValueStrategy. the address must belong to the account.

Storage(String)

Storage access error.

Panic(String)

Panic error.

InvalidMessageId

Invalid message identifier.

InvalidTransactionId

Invalid transaction identifier.

AddressBuildRequiredField(AddressBuildRequiredField)

Address build error: required field not filled.

AccountInitialiseRequiredField(AccountInitialiseRequiredField)

Account initialisation error: required field not filled.

BeeMessage(Error)

Error from bee_message crate.

InvalidBackupFile

Path provided to import_accounts isn't a valid file

InvalidBackupDestination

Backup destination argument is invalid

StorageAdapterNotDefined

the storage adapter isn't set

MnemonicEncode(String)

Mnemonic generation error.

InvalidMnemonic(String)

Invalid mnemonic error

StorageExists

Can't import accounts because the storage already exist

StorageAdapterNotSet(PathBuf)

Storage adapter not defined for the given storage path.

AccountDecrypt(String)

error decrypting stored account using provided encryptionKey

AccountEncrypt(String)

error encrypting stored account using provided encryptionKey

StorageIsEncrypted

Can't use AccountManager API because the storage is encrypted

CannotUseIndexIdentifier

cannot use index to get account - multiple index sequences found (two or more different signer types stored on accounts)

LedgerMiscError

Ledger transport error

LedgerDongleLocked

Dongle Locked

LedgerDeniedByUser

Denied by User

LedgerDeviceNotFound

Ledger Device not found

LedgerEssenceTooLarge

Ledger Essence Too Large

AccountAliasAlreadyExists

Account alias must be unique.

DustError(String)

Dust error, for example not enough balance on an address.