Module iota_wallet::event

The event module.

Structs

BalanceChange
The balance change event payload.

BalanceEvent
The balance change event data.

BalanceEventHandler

ErrorHandler

StrongholdStatusChangeEventHandler

TransactionConfirmationChangeEvent
A transaction-related event data.

TransactionConfirmationChangeEventHandler

TransactionEvent
A transaction-related event data.

TransactionEventHandler

Enums

TransactionEventType

Traits

EventHandler

Functions

add_transaction_listener
Adds a transaction-related event listener.

balance_listeners
Gets the balance change listeners array.

emit_balance_change
Emits a balance change event.

emit_confirmation_state_change
Emits a transaction confirmation state change event.

emit_error

emit_stronghold_status_change

emit_transaction_event
Emits a transaction-related event.

error_listeners
Gets the balance change listeners array.

generate_event_id

on_balance_change
Listen to balance changes.

on_broadcast
Listen to transaction broadcast.

on_confirmation_state_change
Listen to transaction confirmation state change.

on_error
Listen to errors.

on_new_transaction
Listen to new messages.

on_reattachment
Listen to transaction reattachment.

on_stronghold_status_change
Listen to stronghold status change events.

remove_balance_change_listener
Removes the balance change listener associated with the given identifier.

remove_broadcast_listener
Removes the broadcast listener associated with the given identifier.

remove_confirmation_state_change_listener
Removes the new confirmation state change listener associated with the given identifier.

remove_error_listener
Removes the error listener associated with the given identifier.

remove_event_listener

remove_new_transaction_listener
Removes the new transaction listener associated with the given identifier.

remove_reattachment_listener
Removes the reattachment listener associated with the given identifier.

remove_stronghold_status_change_listener
Removes the stronghold status change listener associated with the given identifier.

stronghold_status_change_listeners
Gets the stronghold status change listeners array.

transaction_confirmation_change_listeners
Gets the transaction confirmation change listeners array.

transaction_listeners
Gets the transaction listeners array.

Type Definitions

BalanceListeners

ErrorListeners

EventId
The event identifier type.

StrongholdStatusChangeListeners

TransactionConfirmationChangeListeners

TransactionListeners