Module iota_wallet::message

The message module.

Structs

IndexationPayload

IotaMessage

Message
A message definition.

MessageId

Transfer
A transfer to make a transaction.

TransferBuilder
A transfer to make a transaction.

Value
The transaction Value struct.

Enums

MessageType
Message type.

Output

Payload

RemainderValueStrategy
The strategy to use for the remainder value management when sending funds.

ValueUnit
Possible Value units.

Traits

Packable
A trait to pack and unpack types to and from bytes.